### 2021-04-09 #3160 @whedon
[PRE REVIEW]: microViz: an R package for microbiome data visualization and statistics
**Submitting author:** @david-barnett (<a href="http://orcid.org/0000-0003-1961-7206">David Barnett</a>)
**Repository:** <a href="https://github.com/david-barnett/microViz" target ="_blank">https://github.com/david-barnett/microViz</a>
**Version:** 0.7.0
