### 2021-04-09 #3160 @danielskatz
👋 @david-barnett - You could work on the possibly missing DOI that whedon suggests, but note that this may be incorrect, and remove the prefix from the other. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to the be the first entry in a new comment.
