+++ manuscript/microViz.bib
@@ -41,6 +41,7 @@
 	title = {phyloseq: {An} {R} package for reproducible interactive analysis and graphics of microbiome census data},
 	volume = {8},
 	url = {http://dx.plos.org/10.1371/journal.pone.0061217},
+	doi = {10.1371/journal.pone.0061217},
 	number = {4},
 	journal = {PLoS ONE},
 	author = {McMurdie, Paul J. and Holmes, Susan},
@@ -59,7 +60,7 @@
 	title = {ggtree: an r package for visualization and annotation of phylogenetic trees with their covariates and other associated data},
 	volume = {8},
 	url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.12628},
-	doi = {https://doi.org/10.1111/2041-210X.12628},
+	doi = {10.1111/2041-210X.12628},
 	journal = {Methods in Ecology and Evolution},
 	author = {Yu, Guangchuang and Smith, David K. and Zhu, Huachen and Guan, Yi and Lam, Tommy Tsan-Yuk},
 	year = {2017},
