### 2021-04-11 df1d2d9
ord_explore docs update and rearrange reactive functions internally, but with no functionality changes
