+++ NEWS.md
@@ -1,3 +1,5 @@
+# microViz (development version)
+
 # microViz 0.7.0 - "fickle fixes"
 
 ## Breaking changes
