+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.0
+Version: 0.7.0.9000
 Authors@R: 
     person(given = "David",
            family = "Barnett",
