### 2021-04-11 ed7387c
tax_fix_value factored out
