+++ tests/testthat/test-tax_fix.R
@@ -37,7 +37,6 @@ for (pseq in names(datasets)) {
     )
   })
 
-
   fixed <- suppressWarnings(tax_fix(datasets[[pseq]]))
 
   test_that(paste("tax_fix dataset stays same:", pseq), {
@@ -60,3 +59,15 @@ for (pseq in names(datasets)) {
     )
   }
 }
+
+# tax_common_unknowns ==========================
+test_that("tax_common_unknowns doesn't change", {
+  local_edition(3)
+  for (i in 0:5){
+    expect_snapshot_csv(
+      name = paste0("tax_common_unknowns-", i),
+      object = tax_common_unknowns(min_length = i)
+    )
+  }
+})
+
