+++ vignettes/articles/phyloseq.Rmd
@@ -1,0 +1,164 @@
+---
+title: "Working with phyloseq objects"
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## Intro
+
+This tutorial will show you how microViz makes working with phyloseq objects easier.
+
+```{r setup}
+library(dplyr)
+library(phyloseq)
+library(microViz)
+```
+
+## Getting your data into phyloseq
+
+phyloseq objects are probably the most commonly used data format for working with microbiome data in R.
+
+The creator of phyloseq, [Paul J. McMurdie](https://github.com/joey711), explains the structure of phyloseq objects and how to construct them on the [phyloseq website](https://joey711.github.io/phyloseq/import-data.html#phyloseq-ize_data_already_in_r).
+
+-   [biom format](http://biom-format.org/) files can be imported to phyloseq with the [`import_biom`](https://joey711.github.io/phyloseq/import-data.html#the_import_family_of_functions) function. Such biom files are generated (or can be) from many processing tools including QIIME [1](http://qiime.org/tutorials/index.html) / [2](https://docs.qiime2.org/2021.2/tutorials/exporting/), [MetaPhlAn](http://segatalab.cibio.unitn.it/tools/metaphlan/), and [NG-Tax](https://doi.org/10.3389/fgene.2019.01366).
+
+-   [mothur](https://mothur.org/) output is also directly supported via phyloseq's [`import_mothur`](https://joey711.github.io/phyloseq/import-data.html#_import_mothur) function.
+
+-   [DADA2](https://benjjneb.github.io/dada2/) output can converted to phyloseq according to these [DADA2 handoff instructions](https://benjjneb.github.io/dada2/tutorial.html#bonus-handoff-to-phyloseq)
+
+Don't worry too much about getting all of your sample metadata into your biom file or phyloseq object at the start, as `ps_join()` makes it easy to add sample data later.
+
+```{r example data}
+# This tutorial will just use some example data 
+# It is already available as a phyloseq object from the corncob package
+example_ps <- corncob::ibd_phylo
+example_ps
+```
+
+### Validating your phyloseq
+
+phyloseq checks that your sample and taxa names are consistent across the different slots of the phyloseq object. microViz provides `phyloseq_validate()` to check for and fix other possible problems with your phyloseq that might cause problems in later analyses. It is recommended to run this at the start of your analyses, and fix any problems identified.
+
+```{r validate ps}
+example_ps <- phyloseq_validate(example_ps, remove_undetected = TRUE)
+example_ps <- tax_fix(example_ps)
+
+```
+
+## microViz and phyloseq overview
+
+Once you have a valid phyloseq object, microViz provides several helpful functions for manipulating that object. The names and syntax of some functions will be familiar to users of dplyr, and reading [the dplyr help pages](https://dplyr.tidyverse.org/) may be useful for getting the most out of these functions.
+
+-   Add a data.frame of metadata to the sample_data slot with `ps_join()`
+
+-   Compute new sample_data variables with `ps_mutate()`
+
+-   Subset or reorder variables in sample_data with `ps_select()`
+
+-   Subset samples based on sample_data variables with `ps_filter()`
+
+-   Reorder samples (can be useful for examining sample data or plotting) with:
+
+    -   `ps_reorder()` : manually set sample order
+
+    -   `ps_arrange()` : order samples using sample_data variables
+
+    -   `ps_seriate()` : order samples according to microbiome similarity
+
+-   Remove duplicated/repeated samples with `ps_dedupe()`
+
[…]
