+++ README.md
@@ -1,38 +1,53 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# microViz <a href='https://david-barnett.github.io/microViz/index.html'><img src='man/figures/microViz-smaller.png' align="right" height="139" /></a>
+# microViz <a href='https://david-barnett.github.io/microViz/index.html'><img src="man/figures/microViz-smaller.png" align="right" height="139"/></a>
 
 <!-- badges: start -->
 
 [![R-CMD-check](https://github.com/david-barnett/microViz/workflows/R-CMD-check/badge.svg)](https://github.com/david-barnett/microViz/actions)
 [![codecov](https://codecov.io/gh/david-barnett/microViz/branch/main/graph/badge.svg?token=C1EoVkhnxA)](https://codecov.io/gh/david-barnett/microViz)
+
 <!-- badges: end -->
 
 microViz provides functions for analysis and visualization of microbiome
-sequencing data. These functions are intended to be easy to use (with
-clear documentation) and modular (for extensibility and flexibility).
-microViz wraps, extends and complements popular microbial ecology
-packages like phyloseq, vegan, and microbiome.
+sequencing data. These functions are intended to be easy to use and
+flexible. microViz wraps, extends and complements popular microbial
+ecology packages like phyloseq, vegan, and microbiome.
 
 **See the documentation website for full details and examples:**
 <https://david-barnett.github.io/microViz/>
 
-  - The [“Get
+-   The [“Get
     started”](https://david-barnett.github.io/microViz/articles/articles/microViz.html)
     page, and also [this
     ReadMe](https://david-barnett.github.io/microViz/index.html), show a
     few example analyses.
-  - The
+
+-   The
     [reference](https://david-barnett.github.io/microViz/reference/index.html)
     page lists all functions and links to their documentation and
     examples
-  - Example video of [Interactive Shiny
+
+-   Example video of [Interactive Shiny
     app](https://david-barnett.github.io/microViz/articles/articles/Interactive-Ordination.html)
-    for exploring your ordination plots\!
-  - Other articles pages discuss and give tutorials and more complicated
-    examples (coming soon\!)
-  - The
+    for exploring your ordination plots!
+
+-   Other articles pages give tutorials and more complicated examples
+
+    -   [Working with phyloseq
+        objects](https://david-barnett.github.io/microViz/articles/articles/phyloseq.html)
+
+    -   [Fixing your taxa table with
+        tax\_fix](https://david-barnett.github.io/microViz/articles/articles/tax-fixing.html)
+
+    -   [Visualising taxonomic compositions with customised
[…]
