+++ pkgdown.yml
@@ -5,6 +5,7 @@ articles:
   articles/Interactive-Ordination: Interactive-Ordination.html
   articles/Visualising-compositions: Visualising-compositions.html
   articles/microViz: microViz.html
+  articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
-last_built: 2021-04-11T19:15Z
+last_built: 2021-04-12T10:51Z
 
