### 2021-04-12 8514ec1
more tests for comp_barplot
