+++ tests/testthat/test-comp_barplot.R
@@ -22,6 +22,61 @@ test_that("fixing tax order of comp_barplot works", {
   )
 })
 
+test_that("comp_barplot doesn't change", {
+  local_edition(3)
+  p <- dietswap %>%
+    ps_filter(timepoint == 1) %>%
+    comp_barplot(
+      tax_level = "Genus", n_taxa = 8,
+      sample_order = "aitchison",
+      merge_other = FALSE
+    ) + coord_flip()
+
+  expect_snapshot_csv(
+    name = "unique_taxa_order",
+    object = levels(p$data$unique)
+  )
+  expect_snapshot_csv(
+    name = "SAMPLE_order",
+    object = levels(p$data$SAMPLE)
+  )
+  expect_snapshot(levels(p$data$top))
+
+})
+
+test_that("list of plots generated by group_by works", {
+  local_edition(3)
+  plots <- dietswap %>%
+    ps_filter(timepoint == 1) %>%
+    comp_barplot(
+      tax_level = "Genus",
+      sample_order = "bray",
+      group_by = "bmi_group"
+    )
+  expect_equal(
+    object = names(plots),
+    expected = c("overweight", "lean", "obese")
+  )
+
+  for (p in names(plots)){
+    expect_snapshot_csv(
+      name = paste0("unique_taxa_order-", p),
+      object = levels(plots[[p]]$data$unique)
+    )
+    expect_snapshot_csv(
+      name = paste0("SAMPLE_order-", p),
+      object = levels(plots[[p]]$data$SAMPLE)
+    )
+    expect_snapshot(
+      {
+        print(p)
+        levels(plots[[p]]$data$top)
+      }
+    )
+  }
+
+})
+
 test_that("tt_add_topN_var helper works as expected", {
   local_edition(3)
   tt <- microViz:::tt_add_topN_var(
