+++ tests/testthat/_snaps/comp_barplot.md
@@ -1,3 +1,75 @@
+# comp_barplot doesn't change
+
+    Code
+      levels(p$data$top)
+    Output
+      [1] "other"                               
+      [2] "Allistipes et rel."                  
+      [3] "Clostridium symbiosum et rel."       
+      [4] "Clostridium cellulosi et rel."       
+      [5] "Faecalibacterium prausnitzii et rel."
+      [6] "Prevotella oralis et rel."           
+      [7] "Oscillospira guillermondii et rel."  
+      [8] "Bacteroides vulgatus et rel."        
+      [9] "Prevotella melaninogenica et rel."   
+
+# list of plots generated by group_by works
+
+    Code
+      print(p)
+    Output
+      [1] "overweight"
+    Code
+      levels(plots[[p]]$data$top)
+    Output
+      [1] "other"                               
+      [2] "Allistipes et rel."                  
+      [3] "Clostridium symbiosum et rel."       
+      [4] "Clostridium cellulosi et rel."       
+      [5] "Faecalibacterium prausnitzii et rel."
+      [6] "Prevotella oralis et rel."           
+      [7] "Oscillospira guillermondii et rel."  
+      [8] "Bacteroides vulgatus et rel."        
+      [9] "Prevotella melaninogenica et rel."   
+
+---
+
+    Code
+      print(p)
+    Output
+      [1] "lean"
+    Code
+      levels(plots[[p]]$data$top)
+    Output
+      [1] "other"                               
+      [2] "Allistipes et rel."                  
+      [3] "Clostridium symbiosum et rel."       
+      [4] "Clostridium cellulosi et rel."       
+      [5] "Faecalibacterium prausnitzii et rel."
+      [6] "Prevotella oralis et rel."           
+      [7] "Oscillospira guillermondii et rel."  
+      [8] "Bacteroides vulgatus et rel."        
+      [9] "Prevotella melaninogenica et rel."   
+
+---
+
+    Code
+      print(p)
+    Output
+      [1] "obese"
+    Code
+      levels(plots[[p]]$data$top)
+    Output
+      [1] "other"                               
+      [2] "Allistipes et rel."                  
+      [3] "Clostridium symbiosum et rel."       
+      [4] "Clostridium cellulosi et rel."       
+      [5] "Faecalibacterium prausnitzii et rel."
+      [6] "Prevotella oralis et rel."           
+      [7] "Oscillospira guillermondii et rel."  
+      [8] "Bacteroides vulgatus et rel."        
+      [9] "Prevotella melaninogenica et rel."   
+
 # tt_add_topN_var helper works as expected
 
     Code
