+++ README.md
@@ -208,8 +208,8 @@ aitchison_perm <- dist_permanova(
   variables = "bmi_group + female"
 )
 #> Dropping samples with missings: 2
-#> 2021-04-12 12:46:53 - Starting PERMANOVA with 99 perms with 1 processes
-#> 2021-04-12 12:46:54 - Finished PERMANOVA
+#> 2021-04-12 13:19:52 - Starting PERMANOVA with 99 perms with 1 processes
+#> 2021-04-12 13:19:52 - Finished PERMANOVA
 # view the permanova results
 perm_get(aitchison_perm) %>% as.data.frame()
 #>            Df   SumOfSqs         R2        F Pr(>F)
@@ -231,8 +231,8 @@ your permanova directly using the ord\_plot function with constraints.
 ``` r
 perm2 <- dist_permanova(data = aitchison_dists, variables = c("weight", "female"), seed = 321)
 #> Dropping samples with missings: 2
-#> 2021-04-12 12:46:54 - Starting PERMANOVA with 999 perms with 1 processes
-#> 2021-04-12 12:46:55 - Finished PERMANOVA
+#> 2021-04-12 13:19:52 - Starting PERMANOVA with 999 perms with 1 processes
+#> 2021-04-12 13:19:53 - Finished PERMANOVA
 perm_get(perm2)
 #> Permutation test for adonis under reduced model
 #> Marginal effects of terms
@@ -308,116 +308,116 @@ devtools::session_info()
 #>  date     2021-04-12                  
 #> 
 #> ─ Packages ───────────────────────────────────────────────────────────────────────────────────────
-#>  package        * version    date       lib source        
-#>  ade4             1.7-16     2020-10-28 [1] CRAN (R 4.0.3)
-#>  ape              5.4-1      2020-08-13 [1] CRAN (R 4.0.3)
-#>  Biobase          2.50.0     2020-10-27 [1] Bioconductor  
-#>  BiocGenerics     0.36.0     2020-10-27 [1] Bioconductor  
-#>  biomformat       1.18.0     2020-10-27 [1] Bioconductor  
-#>  Biostrings       2.58.0     2020-10-27 [1] Bioconductor  
-#>  cachem           1.0.4      2021-02-13 [1] CRAN (R 4.0.3)
-#>  Cairo            1.5-12.2   2020-07-07 [1] CRAN (R 4.0.3)
-#>  callr            3.6.0      2021-03-28 [1] CRAN (R 4.0.3)
-#>  circlize         0.4.12     2021-01-08 [1] CRAN (R 4.0.3)
-#>  cli              2.4.0      2021-04-05 [1] CRAN (R 4.0.3)
-#>  clue             0.3-58     2020-12-03 [1] CRAN (R 4.0.3)
-#>  cluster          2.1.0      2019-06-19 [2] CRAN (R 4.0.3)
-#>  codetools        0.2-18     2020-11-04 [1] CRAN (R 4.0.3)
-#>  colorspace       2.0-0      2020-11-11 [1] CRAN (R 4.0.3)
-#>  ComplexHeatmap   2.6.2      2020-11-12 [1] Bioconductor  
-#>  crayon           1.4.1      2021-02-08 [1] CRAN (R 4.0.3)
[…]
