+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.0.9001
+Version: 0.7.1
 Authors@R: 
     person(given = "David",
            family = "Barnett",
