+++ README.Rmd
@@ -70,7 +70,7 @@ if (!requireNamespace("BiocManager", quietly = TRUE))
 BiocManager::install(c("phyloseq", "microbiome"))
 
 # # Installing the latest version of this package # #
-devtools::install_github("david-barnett/microViz@0.7.0") # check 0.7.0 is the latest version?
+devtools::install_github("david-barnett/microViz@0.7.1") # check 0.7.1 is the latest version?
 # advanced tip: add @<commit-hash> after microViz to install a version from a particular commit
 ```
 
