+++ docker/latest/Dockerfile
@@ -1,0 +1,3 @@
+FROM rocker/verse:4.0.4
+
+RUN R -e "remotes::install_github('david-barnett/microViz')"
