### 2021-04-15 f40eb80
rcmdcheck on windows and ubuntu and more R versions
