### 2021-04-15 fd6e2dd
stop testing on R 3.6 several test failures suggest something going wrong with dependencies: possibly microbiome and complexheatmap
