### 2021-04-15 2e535a6
relax test for ord_plot, allow flipped PCoA axes (windows vs nix discrepancy) is still equivalent
