+++ tests/testthat/_snaps/ord_plot.md
@@ -57,16 +57,16 @@
 # aitchison plot hasn't changed
 
     Code
-      cat(p3$data[1:50, 1, drop = TRUE])
+      cat(abs(p3$data[1:50, 1, drop = TRUE]))
     Output
-      -1.106491 0.8693114 0.7769138 0.1722429 0.5494207 1.342983 0.6422256 0.6158025 1.381771 1.076831 1.047452 0.9241626 1.112544 0.6735143 0.9536031 0.7115146 0.4646272 -0.09576095 0.1244906 0.7623212 0.6518461 0.09280437 0.7242988 0.5598014 0.8549104 -0.2034272 0.2821326 1.169105 0.3978237 0.5269234 0.8514695 -1.310387 -1.287183 -0.0458201 0.5668595 -1.1262 -0.6357289 -0.044228 -0.9423226 -0.0271642 -1.315038 -1.145424 -0.1641273 0.2225042 -0.8148921 -0.1271763 -0.3235937 -0.8928878 0.0088699 0.9016179
+      1.106491 0.8693114 0.7769138 0.1722429 0.5494207 1.342983 0.6422256 0.6158025 1.381771 1.076831 1.047452 0.9241626 1.112544 0.6735143 0.9536031 0.7115146 0.4646272 0.09576095 0.1244906 0.7623212 0.6518461 0.09280437 0.7242988 0.5598014 0.8549104 0.2034272 0.2821326 1.169105 0.3978237 0.5269234 0.8514695 1.310387 1.287183 0.0458201 0.5668595 1.1262 0.6357289 0.044228 0.9423226 0.0271642 1.315038 1.145424 0.1641273 0.2225042 0.8148921 0.1271763 0.3235937 0.8928878 0.0088699 0.9016179
 
 ---
 
     Code
-      cat(p3$data[1:50, 2, drop = TRUE])
+      cat(abs(p3$data[1:50, 2, drop = TRUE]))
     Output
-      0.5787119 -1.110654 -1.004009 -1.55686 0.476133 0.3927158 -1.640435 -0.4383189 -0.3547506 -1.092631 -0.5076862 -1.161595 -1.458424 -2.499746 0.4842615 -0.7047841 -0.3034085 -1.131942 -1.505218 -0.7052604 -1.343786 -1.922793 -0.1882586 -0.7064074 -1.607724 -1.576125 -0.8710285 -0.9390623 -1.286664 -2.279908 0.4152985 0.2621039 -0.2757177 -0.1157476 0.2378094 -0.2285831 0.1264438 0.9902263 -0.1533971 -1.291777 0.08253041 -0.191776 0.2938356 0.6350377 1.308276 1.170076 1.064223 0.4547523 -0.06588292 0.6282335
+      0.5787119 1.110654 1.004009 1.55686 0.476133 0.3927158 1.640435 0.4383189 0.3547506 1.092631 0.5076862 1.161595 1.458424 2.499746 0.4842615 0.7047841 0.3034085 1.131942 1.505218 0.7052604 1.343786 1.922793 0.1882586 0.7064074 1.607724 1.576125 0.8710285 0.9390623 1.286664 2.279908 0.4152985 0.2621039 0.2757177 0.1157476 0.2378094 0.2285831 0.1264438 0.9902263 0.1533971 1.291777 0.08253041 0.191776 0.2938356 0.6350377 1.308276 1.170076 1.064223 0.4547523 0.06588292 0.6282335
 
 ---
 
