### 2021-04-21 a33fba8
fix shiny 1.5 incompatibility by explicitly using session
