+++ NEWS.md
@@ -1,3 +1,5 @@
+# microViz (development version)
+
 # microViz 0.7.1
 
 ## Fixes
