### 2021-04-21 6536913
fix session bug shiny1.5 updatesliderinput
