### 2021-04-21 096a28b
full shiny 1.5 compatibility
