+++ README.md
@@ -10,9 +10,9 @@
 
 <!-- badges: end -->
 
-microViz provides functions for analysis and visualization of microbiome
-sequencing data. These functions are intended to be easy to use and
-flexible. microViz wraps, extends and complements popular microbial
+:package: microViz is an R package for analysis and visualization of
+microbiome sequencing data. microViz functions are intended to be easy
+to use and flexible. microViz extends and complements popular microbial
 ecology packages like phyloseq, vegan, and microbiome.
 
 **See the documentation website for full details and examples:**
@@ -66,18 +66,34 @@ following instructions.
 # for installing from github you'll need the devtools package
 install.packages("devtools", type = "binary") # (not binary if you're on linux)
 
-# If you don't already have the latest versions of phyloseq and microbiome, you should install these from bioconductor:
+# If you don't already have the latest versions of phyloseq and microbiome, you can install these from Bioconductor:
 
 if (!requireNamespace("BiocManager", quietly = TRUE))
  install.packages("BiocManager", type = "binary") # (not binary if you're on linux)
 BiocManager::install(c("phyloseq", "microbiome"))
 
-# # Installing the latest version of this package # #
-devtools::install_github("david-barnett/microViz@0.7.1") # check 0.7.1 is the latest version?
+# # Installing the latest "released" version of this package # #
+devtools::install_github("david-barnett/microViz@0.7.1") # check 0.7.1 is the latest release
+
+# # If you encounter bugs: please try installing the very latest development version:
+devtools::install_github("david-barnett/microViz")
+
 # advanced tip: add @<commit-hash> after microViz to install a version from a particular commit
 ```
 
-## Examples below
+:package: I highly recommend using
+[renv](https://rstudio.github.io/renv/index.html) for managing your R
+package installations across multiple projects.
+
+:whale: Alternatively, for Docker users an image with the main branch
+installed is available at:
+<https://hub.docker.com/r/barnettdavid/microviz-rocker-verse>
+
+:date: microViz is tested to work with R version 4.0.0 and higher, on
+Windows, MacOS, and Ubuntu 18 and 20. R version 3.6.\* should probably
+work, but I don’t formally test this.
+
+## A few examples
 
 ``` r
 library(microViz)
@@ -100,7 +116,7 @@ dietswap <- dietswap %>%
 sample_data(dietswap)$female[c(3, 4)] <- NA
 ```
 
-## Looking at your data
+### Looking at your data
 
 You have quite a few samples in your phyloseq object, and would like to
 visualise their compositions. Perhaps these example data differ across
@@ -127,7 +143,7 @@ dietswap %>%
 
 <img src="man/figures/README-unnamed-chunk-2-1.png" width="100%" />
 
-## Example ordination plot workflow
+### Example ordination plot workflow
 
 Maybe visually inspecting all your samples isn’t quite what you want.
 Ordination methods can also help you to visualise if overall microbial
@@ -181,7 +197,7 @@ customised_plot
 
[…]
