+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.1.9001
+Version: 0.7.1.9002
 Authors@R: 
     person(given = "David",
            family = "Barnett",
@@ -30,7 +30,7 @@ Imports:
     purrr,
     tibble,
     tidyr,
-    shiny,
+    shiny (>= 1.5.0),
     scales,
     colorspace,
     tidygraph,
