+++ README.Rmd
@@ -248,7 +248,7 @@ cor_heatmap(psq, taxa, anno_tax = tax_anno(undetected = 50))
 
 :innocent: If you find any part of microViz useful to your work, please consider citing the Zenodo archive DOI: <https://doi.org/10.5281/zenodo.4644057>
 
-Publication pending: microViz has been submitting to a peer-reviewed journal.
+Publication pending: microViz has been submitted to a peer-reviewed journal.
 
 ## Session info
 
