+++ README.Rmd
@@ -21,7 +21,7 @@ options(width = 100)
 
 <!-- badges: start -->
 
-[![R-CMD-check](https://github.com/david-barnett/microViz/workflows/R-CMD-check/badge.svg)](https://github.com/david-barnett/microViz/actions) [![codecov](https://codecov.io/gh/david-barnett/microViz/branch/main/graph/badge.svg?token=C1EoVkhnxA)](https://codecov.io/gh/david-barnett/microViz)
+[![R-CMD-check](https://github.com/david-barnett/microViz/workflows/R-CMD-check/badge.svg)](https://github.com/david-barnett/microViz/actions) [![codecov](https://codecov.io/gh/david-barnett/microViz/branch/main/graph/badge.svg?token=C1EoVkhnxA)](https://codecov.io/gh/david-barnett/microViz) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4680616.svg)](https://doi.org/10.5281/zenodo.4680616) ![Docker Cloud Build Status](https://img.shields.io/docker/cloud/build/barnettdavid/microviz-rocker-verse)
 
 <!-- badges: end -->
 
@@ -70,7 +70,7 @@ if (!requireNamespace("BiocManager", quietly = TRUE))
 BiocManager::install(c("phyloseq", "microbiome"))
 
 # # Installing the latest "released" version of this package # #
-devtools::install_github("david-barnett/microViz@0.7.1") # check 0.7.1 is the latest release
+devtools::install_github("david-barnett/microViz@0.7.2") # check 0.7.2 is the latest release
 
 # # If you encounter bugs: please try installing the very latest development version:
 devtools::install_github("david-barnett/microViz")
