+++ README.md
@@ -7,6 +7,9 @@
 
 [![R-CMD-check](https://github.com/david-barnett/microViz/workflows/R-CMD-check/badge.svg)](https://github.com/david-barnett/microViz/actions)
 [![codecov](https://codecov.io/gh/david-barnett/microViz/branch/main/graph/badge.svg?token=C1EoVkhnxA)](https://codecov.io/gh/david-barnett/microViz)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4680616.svg)](https://doi.org/10.5281/zenodo.4680616)
+![Docker Cloud Build
+Status](https://img.shields.io/docker/cloud/build/barnettdavid/microviz-rocker-verse)
 
 <!-- badges: end -->
 
@@ -73,7 +76,7 @@ if (!requireNamespace("BiocManager", quietly = TRUE))
 BiocManager::install(c("phyloseq", "microbiome"))
 
 # # Installing the latest "released" version of this package # #
-devtools::install_github("david-barnett/microViz@0.7.1") # check 0.7.1 is the latest release
+devtools::install_github("david-barnett/microViz@0.7.2") # check 0.7.2 is the latest release
 
 # # If you encounter bugs: please try installing the very latest development version:
 devtools::install_github("david-barnett/microViz")
@@ -224,8 +227,8 @@ aitchison_perm <- dist_permanova(
   variables = "bmi_group + female"
 )
 #> Dropping samples with missings: 2
-#> 2021-04-21 10:35:54 - Starting PERMANOVA with 99 perms with 1 processes
-#> 2021-04-21 10:35:54 - Finished PERMANOVA
+#> 2021-04-21 21:30:32 - Starting PERMANOVA with 99 perms with 1 processes
+#> 2021-04-21 21:30:32 - Finished PERMANOVA
 # view the permanova results
 perm_get(aitchison_perm) %>% as.data.frame()
 #>            Df   SumOfSqs         R2        F Pr(>F)
@@ -247,8 +250,8 @@ your permanova directly using the ord\_plot function with constraints.
 ``` r
 perm2 <- dist_permanova(data = aitchison_dists, variables = c("weight", "female"), seed = 321)
 #> Dropping samples with missings: 2
-#> 2021-04-21 10:35:54 - Starting PERMANOVA with 999 perms with 1 processes
-#> 2021-04-21 10:35:55 - Finished PERMANOVA
+#> 2021-04-21 21:30:33 - Starting PERMANOVA with 999 perms with 1 processes
+#> 2021-04-21 21:30:33 - Finished PERMANOVA
 perm_get(perm2)
 #> Permutation test for adonis under reduced model
 #> Marginal effects of terms
@@ -337,7 +340,7 @@ devtools::session_info()
 #>  ade4             1.7-16   2020-10-28 [1] CRAN (R 4.0.3)
 #>  ape              5.4-1    2020-08-13 [1] CRAN (R 4.0.3)
 #>  Biobase          2.50.0   2020-10-27 [1] Bioconductor  
-#>  BiocGenerics     0.36.0   2020-10-27 [1] Bioconductor  
+#>  BiocGenerics     0.36.1   2021-04-16 [1] Bioconductor  
 #>  biomformat       1.18.0   2020-10-27 [1] Bioconductor  
 #>  Biostrings       2.58.0   2020-10-27 [1] Bioconductor  
[…]
