### 2021-04-21 #3201 @whedon
[REVIEW]: microViz: an R package for microbiome data visualization and statistics
**Submitting author:** @david-barnett (<a href="http://orcid.org/0000-0003-1961-7206">David Barnett</a>)
**Repository:** <a href="https://github.com/david-barnett/microViz" target ="_blank">https://github.com/david-barnett/microViz</a>
**Version:** v0.7.9

@marypiper, @yoonjeongcha, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
