+++ man/ord_explore.Rd
@@ -54,7 +54,7 @@ library(dplyr)
 data("enterotype")
 
 # simple example #
-taxa_names(enterotype)[1] <- "unclassified"
+taxa_names(enterotype)[1] <- "unclassified" # replaces the "-1" taxon name
 ps <- tax_fix(enterotype) # remove NA taxa
 ord1 <- ps \%>\%
   dist_calc("bray") \%>\%
@@ -62,7 +62,8 @@ ord1 <- ps \%>\%
 
 # ord_explore(data = ord1, auto_caption = NA)
 
-# constrained biplot example #
+# constrained ordination example #
+
 data("dietswap", package = "microbiome")
 
 # create a couple of numerical variables to use as constraints
@@ -77,9 +78,6 @@ constrained_aitchison_rda <- dietswap \%>\%
   tax_transform("clr") \%>\%
   ord_calc(constraints = c("weight", "female"))
 
-# note if you want to visualise an ordination that required transformation
-# you must take care to provide an untransformed phyloseq to ps,
-# otherwise the compositions plotted will be also transformed!
 # constrained_aitchison_rda \%>\%
 #   ord_explore(
 #     plot_taxa = 1:5, tax_lab_style = list(size = 3),
@@ -95,6 +93,7 @@ constrained_aitchison_rda <- dietswap \%>\%
 #   ord_calc() \%>\%
 #   ord_explore()
 
+# another dataset
 data("soilrep", package = "phyloseq")
 # test auto creation of SAMPLE var
 ps <- soilrep \%>\% ps_select(-Sample)
