+++ vignettes/articles/phyloseq.Rmd
@@ -135,7 +135,7 @@ partial_ps %>%
   tax_agg("Genus") %>% 
   ps_arrange(desc(Bacteroides), .target = "otu_table") %>% 
   otu_get() %>% # get the otu table
-  .[, 1:7] # show only a subset of the otu_table
+  .[, 1:6] # show only a subset of the otu_table
 
 # Plot samples' compositions in this order
 partial_ps %>% 
