+++ vignettes/ordination.Rmd
@@ -1,0 +1,122 @@
+---
+title: "Ordination plots"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Ordination plots}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## Intro
+
+This article will show you how to create and customise ordination plots, like PCA and RDA, with microViz.
+
+```{r setup}
+library(phyloseq)
+library(ggplot2)
+library(microViz)
+```
+
+We will use example data from stool samples from an inflammatory bowel disease (IBD) study, borrowed from the great `corncob` package. See the article about [working with phyloseq objects](https://david-barnett.github.io/microViz/articles/articles/phyloseq.html) if you want to get started with your own data, or just to learn more about manipulating phyloseq objects with microViz.
+
+```{r get data}
+ibd <- corncob::ibd_phylo
+ibd
+```
+
+First we check the phyloseq object and fix any uninformative tax_table entries (see the article on [fixing your tax_table](https://david-barnett.github.io/microViz/articles/articles/tax-fixing.html) for more info).
+
+```{r fix it}
+ibd <- phyloseq_validate(ibd, remove_undetected = TRUE)
+ibd <- tax_fix(ibd) # try tax_fix_interactive if you have problems with your own data
+```
+
+## Motivation
+
+Ordination plots are a great way to see any clustering or other patterns of microbiota (dis)similarity in (many) samples. Ordinations like PCA or PCoA show the largest patterns of variation in your data, and constrained ordination techniques like RDA or CCA can show you microbial variation that could be explained by other variables in your sample_data (but interpret constrained ordinations with care, and ideally test for the statistical significance of any hypothesised associations using a method like PERMANOVA, see `dist_permanova()`).
+
+Check out the [Gusta Me website](https://sites.google.com/site/mb3gustame/) for a gentle theoretical introduction to [PCA](https://sites.google.com/site/mb3gustame/indirect-gradient-analysis/pca), [PCoA](https://sites.google.com/site/mb3gustame/dissimilarity-based-methods/principal-coordinates-analysis), [RDA](https://sites.google.com/site/mb3gustame/constrained-analyses/rda), [CCA](https://sites.google.com/site/mb3gustame/constrained-analyses/cca) and more.
+
+Ordination plots can also be paired with barplots for greater insight into microbial compositions, e.g. see `ord_plot_iris()` and the `ord_explore()` interactive [Shiny app](https://david-barnett.github.io/microViz/articles/articles/Interactive-Ordination.html).
+
+## Prepare your microbes
+
+When creating an ordination plot, you first need to prepare the microbiota variables.
+
+-   Decide at which taxonomic rank to aggregate your data, e.g. "Genus"
+
+-   Consider transforming the microbial counts, e.g. using the "clr" (centred log ratio) transformation, which is often recommended for compositional data ([like sequencing data](https://doi.org/10.3389/fmicb.2017.02224 "Gloor 2017"))
+
+```{r clr}
+ibd %>% 
+  tax_transform(transformation = "clr", rank = "Genus")
+```
+
+-   Some methods, such as [PCoA](#pcoa), require a matrix of pairwise distances between samples, which you can easily calculate with `dist_calc()`. Normally you should NOT transform your data when using a distance-based method, but it is useful to record an "identity" transformation anyway, to make it clear you have not transformed your data.
+
+```{r identity}
+ibd %>% 
+  tax_transform(transformation = "identity", rank = "Genus") %>% 
+  dist_calc("bray") # bray curtis distance
+```
+
[…]
