+++ pkgdown.yml
@@ -7,5 +7,6 @@ articles:
   articles/microViz: microViz.html
   articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
-last_built: 2021-04-21T19:53Z
+  ordination: ordination.html
+last_built: 2021-04-22T15:37Z
 
