+++ tests/testthat/test-comp_barplot.R
@@ -22,6 +22,13 @@ test_that("fixing tax order of comp_barplot works", {
   )
 })
 
+top8bysum <- c(
+  "other", "Allistipes et rel.", "Clostridium symbiosum et rel.",
+  "Clostridium cellulosi et rel.", "Faecalibacterium prausnitzii et rel.",
+  "Prevotella oralis et rel.", "Oscillospira guillermondii et rel.",
+  "Bacteroides vulgatus et rel.", "Prevotella melaninogenica et rel."
+)
+
 test_that("comp_barplot doesn't change", {
   local_edition(3)
   p <- dietswap %>%
@@ -29,7 +36,8 @@ test_that("comp_barplot doesn't change", {
     comp_barplot(
       tax_level = "Genus", n_taxa = 8,
       sample_order = "aitchison",
-      merge_other = FALSE
+      merge_other = FALSE,
+      tax_order = sum,
     ) + coord_flip()
 
   expect_snapshot_csv(
@@ -40,8 +48,7 @@ test_that("comp_barplot doesn't change", {
     name = "SAMPLE_order",
     object = levels(p$data$SAMPLE)
   )
-  expect_snapshot(levels(p$data$top))
-
+  expect_equal(object = levels(p$data$top), expected = top8bysum)
 })
 
 test_that("list of plots generated by group_by works", {
@@ -58,7 +65,7 @@ test_that("list of plots generated by group_by works", {
     expected = c("overweight", "lean", "obese")
   )
 
-  for (p in names(plots)){
+  for (p in names(plots)) {
     expect_snapshot_csv(
       name = paste0("unique_taxa_order-", p),
       object = levels(plots[[p]]$data$unique)
@@ -67,14 +74,9 @@ test_that("list of plots generated by group_by works", {
       name = paste0("SAMPLE_order-", p),
       object = levels(plots[[p]]$data$SAMPLE)
     )
-    expect_snapshot(
-      {
-        print(p)
-        levels(plots[[p]]$data$top)
-      }
-    )
+    expect_snapshot(print(p))
+    expect_equal(levels(plots[[p]]$data$top), top8bysum)
   }
-
 })
 
 test_that("tt_add_topN_var helper works as expected", {
