+++ vignettes/ordination.Rmd
@@ -16,12 +16,13 @@ knitr::opts_chunk$set(
 
 ## Intro
 
-This article will show you how to create and customise ordination plots, like PCA and RDA, with microViz.
+This article will show you how to create and customise ordination plots, like [PCA](#pca) and [RDA](#rda), with microViz.
 
 ```{r setup}
 library(phyloseq)
 library(ggplot2)
 library(microViz)
+knitr::opts_chunk$set(fig.width = 7, fig.height = 5)
 ```
 
 We will use example data from stool samples from an inflammatory bowel disease (IBD) study, borrowed from the great `corncob` package. See the article about [working with phyloseq objects](https://david-barnett.github.io/microViz/articles/articles/phyloseq.html) if you want to get started with your own data, or just to learn more about manipulating phyloseq objects with microViz.
@@ -71,15 +72,15 @@ ibd %>%
 
 ### ps_extra
 
-Notice that the objects created above are of class "ps_extra". This is just a simple S3 list object that holds your phyloseq object and additional stuff created from this phyloseq object, such as a distance matrix, as well as info in any transformation and aggregation applied to your taxa. microViz uses this to automatically create plot captions, to help you and your collaborators remember how you made each plot! You can access the phyloseq object, distance matrix and other parts of a ps_extra object with `ps_get()`, `dist_get()`, and friends.
+Notice that the objects created above are of class "ps_extra". This is just a simple S3 list object that holds your phyloseq object and additional stuff created from this phyloseq object, such as a distance matrix, as well as info on any transformation and aggregation applied to your taxa. microViz uses this to automatically create plot captions, to help you and your collaborators remember how you made each plot! You can access the phyloseq object, distance matrix and other parts of a ps_extra object with `ps_get()`, `dist_get()`, and friends.
 
 ## PCA
 
-Principle ***components*** analysis is an unconstrained method that does not use a distance matrix. PCA directly uses the (transformed) microbial variables, so you do not need `dist_calc()`. `ord_calc` performs the ordination (adding it to the ps_extra object) and `ord_plot()` creates the ggplot2 scatterplot (which you can customise like other ggplot objects). 
+Principle ***Components*** Analysis is an unconstrained method that does not use a distance matrix. PCA directly uses the (transformed) microbial variables, so you do not need `dist_calc()`. `ord_calc` performs the ordination (adding it to the ps_extra object) and `ord_plot()` creates the ggplot2 scatterplot (which you can customise like other ggplot objects).
 
 Each point is a sample, and samples that appear closer together are typically more similar to each other than samples which are further apart. So by colouring the points by IBD status you can see that the microbiota from people with IBD is often, but not always, highly distinct from people without IBD.
 
-```{r pca, fig.width=7, fig.height=5}
+```{r pca}
 ibd %>% 
   tax_transform("clr", rank = "Genus") %>% 
   # when no distance matrix or constraints are supplied, PCA is the default/auto ordination method
@@ -92,7 +93,7 @@ One benefit of not using a distance matrix, is that you can plot taxa "loadings"
 
[…]
