+++ docker/release/Dockerfile
@@ -1,0 +1,3 @@
+FROM rocker/verse:latest
+
+RUN R -e "remotes::install_github('david-barnett/microViz', ref = remotes::github_release())"
