+++ _pkgdown.yml
@@ -14,6 +14,7 @@ reference:
   - ps_dedupe
   - ps_drop_incomplete
   - ps_otu2samdat
+  - ps_melt
 
 - title: "Manipulating taxa"
   desc: >
@@ -28,6 +29,10 @@ reference:
   - tax_scale
   - tax_select
   - tax_name
+  - tax_sort
+  - tax_reorder
+  - tax_top
+  - tax_names2tt
 
 - title: "Ordination and related methods"
   desc: >
@@ -55,14 +60,17 @@ reference:
   desc: >
     Annotated heatmaps and stacked barcharts
 - contents:
+  - comp_barplot
   - comp_heatmap
   - cor_heatmap
   - tax_anno
-  - comp_barplot
+  - heat_palette
+  - heat_numbers
+  - heat_grid
 
 - title: "Other functions"
   desc: >
-    Helpful funs, but less interesting...
+    Helpful functions, but less exciting...
 - contents:
   - distinct_palette
   - phyloseq_validate
