+++ man/tax_fill_unknowns.Rd
@@ -1,41 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tax_fill_unknowns_deprecated.R
-\name{tax_fill_unknowns}
-\alias{tax_fill_unknowns}
-\title{Deprecated, use tax_fix}
-\usage{
-tax_fill_unknowns(
-  ps = NULL,
-  min_length = NULL,
-  unknowns = NULL,
-  levels = NULL,
-  suffix_rank = NULL,
-  sep = NULL,
-  anon_unique = NULL,
-  verbose = NULL
-)
-}
-\arguments{
-\item{ps}{deprecated, use tax_fix}
-
-\item{min_length}{deprecated, use tax_fix}
-
-\item{unknowns}{deprecated, use tax_fix}
-
-\item{levels}{deprecated, use tax_fix}
-
-\item{suffix_rank}{deprecated, use tax_fix}
-
-\item{sep}{deprecated, use tax_fix}
-
-\item{anon_unique}{deprecated, use tax_fix}
-
-\item{verbose}{deprecated, use tax_fix}
-}
-\value{
-nothing
-}
-\description{
-tax_fill_unknowns is deprecated, use tax_fix instead (with same arguments)
-or try out tax_fix_interactive()
-}
