+++ NAMESPACE
@@ -43,7 +43,6 @@ export(ps_seriate)
 export(samdat_tbl)
 export(tax_agg)
 export(tax_anno)
-export(tax_fill_unknowns)
 export(tax_filter)
 export(tax_fix)
 export(tax_fix_interactive)
