+++ tests/testthat/test-tax_fix_interactive.R
@@ -3,7 +3,6 @@ library(shinytest)
 test_that("tax_fix_interactive() works", {
   testthat::skip_on_cran()
   testthat::skip_on_bioc()
-  testthat::skip_on_ci()
 
   expect_pass(testApp(appDir = "apps/tax_fix_interactive/", compareImages = FALSE))
 })
