+++ tests/testthat/test-ord_explore.R
@@ -3,7 +3,8 @@ library(shinytest)
 test_that("ord_explore() works", {
   testthat::skip_on_cran()
   testthat::skip_on_bioc()
-  testthat::skip_on_ci()
+  testthat::skip_on_os("windows")
+  testthat::skip_on_os("mac")
 
   expect_pass(testApp(appDir = "apps/ord_explore/", compareImages = FALSE))
 })
