+++ tests/testthat/apps/tax_fix_interactive/tests/shinytest/mytest.R
@@ -14,3 +14,10 @@ app$setInputs(sep = " ...")
 app$setInputs(selected = character(0))
 app$setInputs(selected = "Anaerofustis")
 app$snapshot()
+
+# https://github.com/rfaelens/exampleShinyTest/blob/master/tests/shinyTest/tests/test.R
+# wait for the process to close gracefully
+# this allows covr to write out the coverage results
+p <- app$.__enclos_env__$private$shinyProcess
+p$interrupt()
+p$wait()
