### 2021-04-24 fd3bd5f
improved tests for tax_fix_interactive
