+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.3.9001
+Version: 0.7.3.9002
 Authors@R: 
     person(given = "David",
            family = "Barnett",
@@ -20,13 +20,11 @@ Imports:
     vegan,
     phyloseq,
     dplyr (>= 1.0.0),
-    GUniFrac,
     rlang (>= 0.4.0),
     ggplot2,
     seriation,
     ggraph (>= 2.0.0),
     future.apply,
-    future,
     purrr,
     tibble,
     tidyr,
@@ -42,6 +40,8 @@ Imports:
     ggiraph,
     DT
 Suggests: 
+    GUniFrac,
+    future,
     stringr,
     patchwork (>= 1.0.0),
     cowplot,
