### 2021-04-24 65ea1af
tidy up tax_model code, mostly just shorter lines
