+++ R/tax_model.R
@@ -2,18 +2,23 @@
 #' @title Statistical modelling for individual taxa in a phyloseq
 #'
 #' @description
-#' `tax_model` provides a simple framework to statistically model the abundance of individual taxa in your data.
-#' You can choose which type of statistical model you want to fit, and you can choose at which rank and (optionally) which specific taxa to fit statistical models for.
+#' `tax_model` provides a simple framework to statistically model the abundance
+#' of individual taxa in your data.
+#' You can choose which type of statistical model you want to fit, and you can
+#' choose at which rank and (optionally) which specific taxa to fit statistical models for.
 #' `tax_model` takes a phyloseq and returns a list of statistical models, one model for each taxon.
-#' The same independent variables are used for all models, as specified in `variables` or `formula` argument (latter takes precedence).
+#' The same independent variables are used for all models,
+#' as specified in `variables` or `formula` argument (latter takes precedence).
 #'
-#' `taxatree_models` runs `tax_model` on every taxon at multiple taxonomic ranks (you choose which ranks with the plural `tax_levels` argument),
+#' `taxatree_models` runs `tax_model` on every taxon at multiple taxonomic ranks
+#' (you choose which ranks with the plural `tax_levels` argument),
 #' and returns the results as a named nested list designed for use with `taxatree_plots`.
 #' One list per rank, one model per taxon at each rank.
 #'
 #'
 #' `type = "bbdml"` will run beta binomial regression model(s) using the `corncob` package.
-#' For bbdml the same formula/variables is/are used for modelling both the abundance and dispersion parameters.
+#' For bbdml the same formula/variables is/are used for modelling both the
+#' abundance and dispersion parameters.
 #'
 #' @details
 #' `tax_model` and `taxatree_models` can use parallel processing with the `future` package.
@@ -79,7 +84,14 @@
 #'   tax_model(tax_level = "Genus", taxa = 1:3, variables = VARS, type = "lm")
 #' @rdname Taxon-modelling
 #' @export
-tax_model <- function(ps, tax_level, type = "bbdml", variables = NULL, formula = NULL, taxa = NULL, verbose = TRUE, ...) {
+tax_model <- function(ps,
+                      tax_level,
+                      type = "bbdml",
+                      variables = NULL,
+                      formula = NULL,
+                      taxa = NULL,
+                      verbose = TRUE,
+                      ...) {
   ps <- ps_get(ps)
   # check phyloseq for common problems (and fix or message about this)
   ps <- phyloseq_validate(ps, remove_undetected = TRUE, verbose = TRUE)
@@ -101,7 +113,7 @@ tax_model <- function(ps, tax_level, type = "bbdml", variables = NULL, formula =
     tt_names <- phyloseq::taxa_names(ps)[taxa]
   } else if (any(!taxa %in% possible_taxa)) {
     stop(
-      "The following taxon names are not present at the chosen aggregation level rank:\n",
+      "The following taxon names are not present at this aggregation rank:\n",
       paste(taxa[!taxa %in% possible_taxa], collapse = "\n")
     )
   } else {
@@ -114,7 +126,10 @@ tax_model <- function(ps, tax_level, type = "bbdml", variables = NULL, formula =
   not_matching <- tt_names != taxons
   if (any(not_matching)) {
     if (!isFALSE(verbose)) {
-      message("Changing ", sum(not_matching), " taxa_names that don't match taxa found at level of ", tax_level)
+      message(
+        "Changing ", sum(not_matching),
[…]
