+++ R/ord_explore.R
@@ -96,6 +96,9 @@ ord_explore <- function(data,
                         ...) {
   # SETUP -------------------------------------------------------------------
 
+  # widths of plots including space for legends, in inches
+  p_width <- c(7, 10)
+
   # create a SAMPLE id variable
   data$ps <- ps_mutate(data$ps, SAMPLE = phyloseq::sample_names(data$ps))
 
@@ -137,7 +140,7 @@ ord_explore <- function(data,
         shiny::tags$style(
           shiny::HTML(
             "body {
-              line-height: 1.2;
+              line-height: 1;
             }
             /* allow dropdown menus to show in split layout input blocks */
             .shiny-split-layout > div {
@@ -165,6 +168,19 @@ ord_explore <- function(data,
                 selected = c(sample_id, "SAMPLE")[[1]]
               )
             ),
+            # shape
+            shiny::splitLayout(
+              cellWidths = c("30%", "70%"),
+              shiny::helpText("Shape:"),
+              shiny::selectInput(
+                inputId = "ord_shape", label = NULL,
+                choices = list(
+                  Variable = phyloseq::sample_variables(ps),
+                  Fixed = ggplot2_shapes()
+                ),
+                selected = "circle filled"
+              )
+            ),
             shiny::splitLayout(
               cellWidths = c("30%", "70%"),
               shiny::helpText("Color:"),
@@ -177,30 +193,12 @@ ord_explore <- function(data,
                 selected = "gray"
               )
             ),
-            shiny::sliderInput(
-              inputId = "ord_alpha", label = "Point alpha",
-              value = 0.6, min = 0, max = 1, ticks = FALSE
-            ),
-            # shape
-            shiny::radioButtons(
-              inputId = "shape_var_type",
-              label = "Point shape:", inline = TRUE,
-              choices = c(
-                "Fixed" = "fixed",
-                "Variable" = "variable"
-              ),
-              selected = "fixed"
-            ),
             shiny::splitLayout(
-              cellWidths = c("35%", "65%"),
-              shiny::numericInput(
-                inputId = "ord_shape_num", label = NULL,
-                value = 21, min = 1, step = 1, max = 21
-              ),
-              shiny::selectInput(
-                inputId = "ord_shape_var", label = NULL,
-                choices = categorical_vars,
-                selected = NULL
+              cellWidths = c("30%", "70%"),
+              shiny::helpText("Alpha:"),
+              shiny::sliderInput(
+                inputId = "ord_alpha", label = NULL,
+                value = 0.6, min = 0, max = 1, ticks = FALSE
               )
             ),
             # size
@@ -303,7 +301,8 @@ ord_explore <- function(data,
               width = 11,
               # interactive ordination plot
               ggiraph::girafeOutput(
-                outputId = "ord_plot", height = "4.5in", width = "6in"
+                outputId = "ord_plot", height = "4.5in",
+                width = paste0(p_width[[1]], "in")
               )
             )
           ),
@@ -315,13 +314,15 @@ ord_explore <- function(data,
                 shiny::tabPanel(
                   title = "ggplot",
                   shiny::plotOutput(
-                    height = "5in", width = "6in", outputId = "comps_gg"
+                    outputId = "comps_gg",
+                    height = "5in", width = paste0(p_width[[2]], "in")
                   )
                 ),
                 shiny::tabPanel(
[…]
