+++ DESCRIPTION
@@ -1,13 +1,13 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.3.9003
+Version: 0.7.3.9004
 Authors@R: 
     person(given = "David",
            family = "Barnett",
            role = c("aut", "cre"),
            email = "david.barnett@maastrichtuniversity.nl",
            comment = c(ORCID = "0000-0003-1961-7206"))
-URL: https://github.com/david-barnett/microViz
+URL: https://david-barnett.github.io/microViz, https://github.com/david-barnett/microViz
 Description: Extends phyloseq and microbiome with further visualization and data manipulation functionality.
 License: GPL (>= 3)
 Encoding: UTF-8
@@ -38,13 +38,13 @@ Imports:
     ComplexHeatmap (>= 2.0.0),
     circlize,
     ggiraph,
+    cowplot,
     DT
 Suggests: 
     GUniFrac,
     future,
     stringr,
     patchwork (>= 1.0.0),
-    cowplot,
     corncob,
     devtools,
     testthat,
