### 2021-04-27 646e9dc
ord_explore conditional panels and safe failure
