+++ R/ord_explore.R
@@ -1,12 +1,21 @@
 #' Interactively explore microbial compositions of ordinated samples
 #'
 #' @description
-#' A Shiny app used like an interactive version of ord_plot (taking the output of ord_calc).
+#' A Shiny app used like an interactive version of `ord_plot()`.
 #' You can select samples on an ordination plot and view their composition with stacked barplots.
 #'
-#' Once running: click and drag to select 2 or more samples to view their compositions.
-#' You can style the ordination plot points and compositional barplot using the options in the left panel.
+#' You can give the `ord_explore()` data argument either of the following:
 #'
+#' - the output of `ord_calc()` (i.e. a ps_extra with an ordination)
+#' - a plain phyloseq object: `ord_explore()` will help you build an ordination
+#'
+#'
+#' Once running:
+#'
+#' 1. edit the ordination if required
+#' 2. style the ordination plot (e.g. choose dimensions; set colour and size; ...)
+#' 3. click or use the lasso tool to select 2 or more samples to view their compositions
+#' 4. style the taxonomic compositions barplot
 #'
 #' @details
 #' If you get an interactive error like the one below:
@@ -20,13 +29,16 @@
 #' that grouping variable cannot be used to select those samples
 #'
 #'
-#' @param data ps_extra list output of ord_calc
+#' @param data ps_extra list output of ord_calc, or phyloseq
 #' @param sample_id name of sample ID variable to use as default for selecting samples
 #' @param seriate_method
 #' seriation method to order phyloseq samples by similarity
 #' @param tax_transform_for_ordering
 #' transform taxa before ordering with ps_seriate
 #' @param app_options passed to shinyApp() options argument
+#' @param plot_widths
+#' widths of plots in inches, including any legends
+#' (first number is ordination, second is composition barplot)
 #' @param ... additional arguments passed to ord_plot
 #'
 #' @return nothing, opens default browser
@@ -36,7 +48,13 @@
 #' library(phyloseq)
 #' library(dplyr)
 #'
-#' # simple example #
+#' if (interactive()) {
+#'   corncob::ibd_phylo %>%
+#'     tax_fix() %>%
+#'     ord_explore()
+#' }
+#'
+#' # simple example with precalculated ordination #
 #' if (interactive()) {
 #'   data("enterotype")
 #'   taxa_names(enterotype)[1] <- "unclassified" # replaces the "-1" taxon name
@@ -46,7 +64,7 @@
 #'     dist_calc("bray") %>%
 #'     ord_calc(method = "PCoA")
 #'
-#'   ord_explore(data = ord1, auto_caption = NA)
+#'   ord_explore(data = ord1, auto_caption = 6)
 #' }
 #'
 #' # constrained ordination example #
@@ -101,11 +119,18 @@ ord_explore <- function(data,
                         seriate_method = "OLO_ward", # ordering samples
                         tax_transform_for_ordering = "identity", # samples
                         app_options = list(launch.browser = TRUE), # shinyApp()
+                        plot_widths = c(6, 7),
                         ...) {
   # SETUP -------------------------------------------------------------------
 
   # widths of plots including space for legends, in inches
-  p_width <- c(7, 9)
+  # TODO how to better handle this choice?
+  p_width <- plot_widths # 1st is ordination, 2nd is composition
+
+  # if data is plain phyloseq, validate and convert to ps_extra
+  if (methods::is(data, "phyloseq")) {
+    data <- tax_transform(phyloseq_validate(data), "identity", rank = "unique")
+  }
 
   # create a SAMPLE id variable
   data$ps <- ps_mutate(data$ps, SAMPLE = phyloseq::sample_names(data$ps))
[…]
