+++ R/ord_explore.R
@@ -48,6 +48,7 @@
 #' library(phyloseq)
 #' library(dplyr)
 #'
+#' # example of quickstart approach with interactive ordination calculation #
 #' if (interactive()) {
 #'   corncob::ibd_phylo %>%
 #'     tax_fix() %>%
@@ -119,7 +120,7 @@ ord_explore <- function(data,
                         seriate_method = "OLO_ward", # ordering samples
                         tax_transform_for_ordering = "identity", # samples
                         app_options = list(launch.browser = TRUE), # shinyApp()
-                        plot_widths = c(6, 7),
+                        plot_widths = c(7, 9),
                         ...) {
   # SETUP -------------------------------------------------------------------
 
@@ -284,10 +285,10 @@ ord_explore <- function(data,
               shiny::splitLayout(
                 cellWidths = c("30%", "65%"),
                 shiny::helpText("Alpha:"),
-                shiny::selectInput(
+                shiny::selectizeInput(
                   inputId = "ord_alpha_var", label = NULL,
-                  choices = numerical_vars,
-                  selected = NULL
+                  choices = numerical_vars, selected = NULL,
+                  options = list(placeholder = 'numeric var?')
                 )
               )
             ),
@@ -310,10 +311,10 @@ ord_explore <- function(data,
               shiny::splitLayout(
                 cellWidths = c("30%", "65%"),
                 shiny::helpText("Size:"),
-                shiny::selectInput(
+                shiny::selectizeInput(
                   inputId = "ord_size_var", label = NULL,
-                  choices = numerical_vars,
-                  selected = NULL
+                  choices = numerical_vars, selected = NULL,
+                  options = list(placeholder = 'numeric var?')
                 )
               )
             )
@@ -391,7 +392,7 @@ ord_explore <- function(data,
                 shiny::helpText("N Distinct:"),
                 shiny::sliderInput(
                   inputId = "taxmax", label = NULL,
-                  min = 1, max = 500, value = 200, step = 1,
+                  min = 1, max = 500, value = 100, step = 1,
                   round = TRUE, ticks = FALSE
                 )
               )
@@ -476,39 +477,35 @@ ord_explore <- function(data,
         shiny::hr(),
         shiny::selectizeInput(
           inputId = "method", label = "Ordination method",
-          choices = c(
-            "auto (select options below!)" = "auto",
-            "PCA (Principle Components Analysis)" = "PCA",
-            "PCoA (Principle Co-ordinates Analysis)" = "PCoA",
-            "RDA (Redundancy Analysis)" = "RDA",
-            "CAP (Constrained PCoA)" = "CAP",
-            "NMDS (Non-metric MDS)" = "NMDS",
-            "CCA (Canonical Correspondence Analysis)" = "CCA"
+          choices = m_choices$ordInfo, selected = m_sel$ordInfo
+        ),
+        shiny::checkboxInput(
+          inputId = "concons",
+          label = "Constrain or condition ordination? (requires numeric vars)",
+          value = FALSE
+        ),
+        shiny::conditionalPanel(
+          condition = "input.concons == true",
+          shiny::selectizeInput(
+            inputId = "const", label = "Constraints", multiple = TRUE,
+            choices = m_choices$const, selected = m_sel$const
           ),
-          selected = m_sel$ordInfo
+          shiny::selectizeInput(
+            inputId = "conds", label = "Conditions", multiple = TRUE,
+            choices = m_choices$conds, selected = m_sel$conds
+          )
         ),
         shiny::selectizeInput(
-          inputId = "rank", label = "Rank",
[…]
