### 2021-04-28 70e6913
ord_explore constraints and conditions
