+++ R/ord_explore.R
@@ -288,7 +288,7 @@ ord_explore <- function(data,
                 shiny::selectizeInput(
                   inputId = "ord_alpha_var", label = NULL,
                   choices = numerical_vars, selected = NULL,
-                  options = list(placeholder = 'numeric var?')
+                  options = list(placeholder = "numeric var?")
                 )
               )
             ),
@@ -314,7 +314,7 @@ ord_explore <- function(data,
                 shiny::selectizeInput(
                   inputId = "ord_size_var", label = NULL,
                   choices = numerical_vars, selected = NULL,
-                  options = list(placeholder = 'numeric var?')
+                  options = list(placeholder = "numeric var?")
                 )
               )
             )
@@ -476,13 +476,21 @@ ord_explore <- function(data,
         shiny::helpText("Choose options to modify ordination created:"),
         shiny::hr(),
         shiny::selectizeInput(
-          inputId = "method", label = "Ordination method",
-          choices = m_choices$ordInfo, selected = m_sel$ordInfo
+          inputId = "rank", label = "Taxonomic Rank",
+          selected = m_sel$rank, choices = m_choices$rank
+        ),
+        shiny::selectizeInput(
+          inputId = "trans", label = "Taxa transformation",
+          choices = m_choices$trans, selected = m_sel$trans
+        ),
+        shiny::selectizeInput(
+          inputId = "dist", label = "Distance / Dissimilarity",
+          choices = m_choices$distInfo, selected = m_sel$distInfo
         ),
         shiny::checkboxInput(
           inputId = "concons",
           label = "Constrain or condition ordination? (requires numeric vars)",
-          value = FALSE
+          value = m_sel$concons
         ),
         shiny::conditionalPanel(
           condition = "input.concons == true",
@@ -496,16 +504,8 @@ ord_explore <- function(data,
           )
         ),
         shiny::selectizeInput(
-          inputId = "dist", label = "Distance / Dissimilarity",
-          choices = m_choices$distInfo, selected = m_sel$distInfo
-        ),
-        shiny::selectizeInput(
-          inputId = "rank", label = "Taxonomic Rank",
-          selected = m_sel$rank, choices = m_choices$rank
-        ),
-        shiny::selectizeInput(
-          inputId = "trans", label = "Taxa transformation",
-          choices = m_choices$trans, selected = m_sel$trans
+          inputId = "method", label = "Ordination method",
+          choices = m_choices$ordInfo, selected = m_sel$ordInfo
         ),
         footer = shiny::tagList(
           shiny::modalButton("Cancel", icon = shiny::icon("times")),
@@ -548,8 +548,10 @@ ord_explore <- function(data,
       # scale = if (is.na(info$scale)) "neither" else info$scale,
       distInfo = if (is.na(info$dist)) "none" else info$dist,
       ordInfo = if (is.na(info$ord)) "auto" else info$ord,
-      const = if (is.na(info$constraints)) NULL else info$constraints,
-      conds = if (is.na(info$conditions)) NULL else info$conditions
+      const = read_cons(info$constraints),
+      conds = read_cons(info$conditions),
+      concons = # constrained or conditioned (checkbox)
+      if (is.na(info$constraints) & is.na(info$conditions)) FALSE else TRUE
     )
 
     #### update selected ------------------------------------------------------
@@ -566,6 +568,7 @@ ord_explore <- function(data,
         # TODO constraints and conditions inputs
         m_sel$const <- input$const
         m_sel$conds <- input$conds
+        m_sel$concons <- input$concons
       }
     )
 
@@ -586,6 +589,33 @@ ord_explore <- function(data,
     )
 
     #### update choices -------------------------------------------------------
[…]
