### 2021-04-28 43c23cd
ord_explore_init helper init
