+++ R/ord_explore.R
@@ -128,45 +128,7 @@ ord_explore <- function(data,
   # TODO how to better handle this choice?
   p_width <- plot_widths # 1st is ordination, 2nd is composition
 
-  # if data is plain phyloseq, validate and convert to ps_extra
-  if (methods::is(data, "phyloseq")) {
-    data <- tax_transform(phyloseq_validate(data), "identity", rank = "unique")
-  }
-
-  # create a SAMPLE id variable
-  data$ps <- ps_mutate(data$ps, SAMPLE = phyloseq::sample_names(data$ps))
-
-  # get info about input data to initialise settings modal choices
-  info <- list(
-    rank = info_get(data)[["tax_agg"]],
-    trans = info_get(data)[["tax_transform"]],
-    scale = info_get(data)[["tax_scale"]],
-    dist = info_get(data)[["distMethod"]],
-    ord = info_get(data)[["ordMethod"]],
-    constraints = info_get(data)[["constraints"]],
-    conditions = info_get(data)[["conditions"]]
-  )
-  # TODO fix ps_seriate so this info isn't necessary
-  if (is.na(info$rank) || is.na(info$trans)) {
-    warning(
-      "Guessing tax rank is 'unique', and transformation is 'identity'",
-      "\nPlease use tax_transform and/or tax_agg to set this info explicitly!"
-    )
-    warnOnStart <- TRUE
-    info$rank <- "unique"
-    info$trans <- "identity"
-  } else {
-    warnOnStart <- FALSE
-  }
-
-  # get list of certain types of variables for populating selectize lists
-  ps <- ps_get(data)
-  samdat <- methods::as(phyloseq::sample_data(ps), "data.frame")
-  numerical_vars <- colnames(
-    samdat[, sapply(X = samdat, function(x) !is.character(x) & !is.factor(x))]
-  )
-  categorical_vars <-
-    colnames(samdat[, sapply(samdat, function(x) !is.numeric(x))])
+  init <- ord_explore_init(data)
 
   message("To stop the app: Click red stop button or hit Esc in the console")
 
@@ -207,16 +169,14 @@ ord_explore <- function(data,
               style = "display:inline-block; width:47.5%",
               shiny::actionButton(
                 inputId = "settings", icon = shiny::icon("cog"),
-                class = "btn-secondary", width = "100%",
-                label = "Edit"
+                label = "Edit", class = "btn-secondary", width = "100%"
               )
             ),
             shiny::div(
               style = "display:inline-block; width:47.5%",
               shiny::actionButton(
                 inputId = "code", icon = shiny::icon("code"),
-                class = "btn-secondary", width = "100%",
-                label = "Code"
+                label = "Code", class = "btn-secondary", width = "100%"
               )
             ),
             shiny::h4("Ordination options"),
@@ -236,21 +196,19 @@ ord_explore <- function(data,
               cellWidths = c("30%", "65%"),
               shiny::helpText("Select:"),
               shiny::selectInput(
-                inputId = "id_var", label = NULL,
-                choices = union("SAMPLE", colnames(samdat)),
-                selected = c(sample_id, "SAMPLE")[[1]]
+                inputId = "id_var", label = NULL, choices = init$vars$all,
+                selected = c(sample_id, "SAMPLE")[[1]] # 'SAMPLE' if id = NULL
               )
             ),
             shiny::splitLayout(
               cellWidths = c("30%", "65%"),
               shiny::helpText("Colour:"),
               shiny::selectInput(
-                inputId = "ord_colour", label = NULL,
+                inputId = "ord_colour", label = NULL, selected = "azure4",
                 choices = list(
-                  Variable = phyloseq::sample_variables(ps),
+                  Variable = init$vars$all,
[…]
