### 2021-04-28 bdbcd15
tweak ps_extra_info print method
