+++ tests/testthat/_snaps/tax_fix.md
@@ -10,7 +10,7 @@
       sample_data() Sample Data:       [ 280 samples by 9 sample variables ]
       tax_table()   Taxonomy Table:    [ 553 taxa by 1 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Genus tax_transform = NA
 
 # tax_fix defaults allow agg: modified_dietswap Phylum
@@ -25,7 +25,7 @@
       sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
       tax_table()   Taxonomy Table:    [ 12 taxa by 1 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Phylum tax_transform = NA
 
 # tax_fix defaults allow agg: modified_dietswap Family
@@ -40,7 +40,7 @@
       sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
       tax_table()   Taxonomy Table:    [ 28 taxa by 2 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Family tax_transform = NA
 
 # tax_fix defaults allow agg: modified_dietswap Genus
@@ -55,7 +55,7 @@
       sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
       tax_table()   Taxonomy Table:    [ 110 taxa by 3 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Genus tax_transform = NA
 
 # tax_fix defaults allow agg: ibd_phylo Kingdom
@@ -70,7 +70,7 @@
       sample_data() Sample Data:       [ 91 samples by 15 sample variables ]
       tax_table()   Taxonomy Table:    [ 1 taxa by 1 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Kingdom tax_transform = NA
 
 # tax_fix defaults allow agg: ibd_phylo Phylum
@@ -85,7 +85,7 @@
       sample_data() Sample Data:       [ 91 samples by 15 sample variables ]
       tax_table()   Taxonomy Table:    [ 9 taxa by 2 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Phylum tax_transform = NA
 
 # tax_fix defaults allow agg: ibd_phylo Class
@@ -100,7 +100,7 @@
       sample_data() Sample Data:       [ 91 samples by 15 sample variables ]
       tax_table()   Taxonomy Table:    [ 19 taxa by 3 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Class tax_transform = NA
 
 # tax_fix defaults allow agg: ibd_phylo Order
@@ -115,7 +115,7 @@
       sample_data() Sample Data:       [ 91 samples by 15 sample variables ]
       tax_table()   Taxonomy Table:    [ 32 taxa by 4 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Order tax_transform = NA
 
 # tax_fix defaults allow agg: ibd_phylo Family
@@ -130,7 +130,7 @@
       sample_data() Sample Data:       [ 91 samples by 15 sample variables ]
       tax_table()   Taxonomy Table:    [ 59 taxa by 5 taxonomic ranks ]
       
-      phyloseq info:
+      ps_extra info:
       tax_agg = Family tax_transform = NA
 
 # tax_fix defaults allow agg: ibd_phylo Genus
@@ -145,7 +145,7 @@
       sample_data() Sample Data:       [ 91 samples by 15 sample variables ]
[…]
