### 2021-04-28 0d1a284
ord_explore dist_choices helper and other tweaks and app smoketest update
