+++ tests/testthat/apps/ord_explore/tests/shinytest/mytest.R
@@ -1,4 +1,0 @@
-app <- ShinyDriver$new("../../")
-app$snapshotInit("mytest")
-
-app$snapshot(list(output = "comps_gg"))
