+++ tests/testthat/apps/ord_explore/shinytest.R
@@ -1,2 +1,0 @@
-library(shinytest)
-shinytest::testApp("../")
