+++ tests/testthat/apps/ord_explore/app.R
@@ -7,6 +7,7 @@ ps <- dietswap %>%
   tax_fix()
 
 ord1 <- ps %>%
+  tax_transform(rank = "unique", transformation = "identity") %>%
   dist_calc("bray") %>%
   ord_calc("PCoA")
 
