+++ R/ord_explore.R
@@ -93,10 +93,13 @@
 #' # try changing the point colour to bmi_group or similar
 #' # (style interactively! e.g. colour doesn't work as argument to ord_explore)
 #'
-#' # another dataset
+#' # another dataset, where "size" variable drives gradient on PC1
+#' # (try setting size and/or alpha to correspond to "size"!)
+#' # then edit the ordination to use "size" as a condition, see what happens
 #' if (interactive()) {
 #'   microbiomeutilities::hmp2 %>%
 #'     tax_fix() %>%
+#'     tax_transform(rank = "Genus", "identity") %>%
 #'     dist_calc("aitchison") %>%
 #'     ord_calc() %>%
 #'     ord_explore()
@@ -216,8 +219,7 @@ ord_explore <- function(data,
               cellWidths = c("30%", "65%"),
               shiny::helpText("Shape:"),
               shiny::selectInput(
-                inputId = "ord_shape", label = NULL,
-                selected = "circle filled",
+                inputId = "ord_shape", label = NULL, selected = "circle",
                 choices = list(
                   Variable = init$vars$all,
                   Fixed = ggplot2_shapes()
@@ -302,8 +304,8 @@ ord_explore <- function(data,
               cellWidths = c("30%", "65%"),
               shiny::helpText("Rank:"),
               shiny::selectInput(
-                inputId = "tax_level_comp", label = NULL, choices = init$ranks,
-                selected = utils::tail(setdiff(init$ranks, "unique"), n = 1)
+                inputId = "tax_level_comp", label = NULL,
+                choices = init$ranks, selected = init$info$rank
               )
             ),
             shiny::splitLayout(
@@ -502,10 +504,13 @@ ord_explore <- function(data,
         m_sel$ordInfo <- input$method
         # TODO scale inputs
         # m_sel$scale <- input$scale
-        # TODO constraints and conditions inputs
         m_sel$const <- input$const
         m_sel$conds <- input$conds
         m_sel$concons <- input$concons
+        # update default choice of taxonomic rank for composition plot
+        shiny::updateSelectizeInput(
+          session = session, inputId = "tax_level_comp", selected = input$rank
+        )
       }
     )
 
@@ -514,17 +519,14 @@ ord_explore <- function(data,
     #### initialise choices ---------------------------------------------------
     m_choices <- shiny::reactiveValues(
       rank = rev(init$ranks), trans = trans_choices(type = "all"),
-      # scale = if (is.na(info$scale)) "neither" else info$scale,
-      distInfo = union(
-        c("none", "bray", "aitchison", "euclidean", "gunifrac"),
-        unlist(phyloseq::distanceMethodList)
-      ),
-      # TODO dist_choices function (separate unifrac and check for phy_tree)
+      # scale = # TODO
+      distInfo = dist_choices(init$data, type = "all"),
       ordInfo = ord_choices(type = "all"),
       const = init$vars$num, conds = init$vars$num
     )
 
     #### update choices -------------------------------------------------------
+
     # modify ordination choices if distance or constraints/conds change
     shiny::observeEvent(
       ignoreInit = TRUE,
@@ -549,6 +551,18 @@ ord_explore <- function(data,
           }
         updateSelectizeInput(
           session = session, inputId = "method", choices = x
+        )
+      }
+    )
+    # remove aitchison distance if log10p or clr transform used
+    shiny::observeEvent(
+      ignoreInit = TRUE,
[…]
