### 2021-04-28 43d5035
rcmdcheck fixes
