+++ pkgdown.yml
@@ -8,5 +8,5 @@ articles:
   articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
   ordination: ordination.html
-last_built: 2021-04-28T20:35Z
+last_built: 2021-04-28T20:45Z
 
