+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.3.9012
+Version: 0.7.3.9013
 Authors@R: 
     person(given = "David",
            family = "Barnett",
