### 2021-04-29 910bd45
rcmdcheck fixes
