### 2021-04-29 2fd1ca5
Clean up ord_explore example
