### 2021-04-29 30e8f1c
fix: ord_plot and ord_explore no longer limited to 6th dimension
