+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.3.9014
+Version: 0.7.3.9015
 Authors@R: 
     person(given = "David",
            family = "Barnett",
