### 2021-05-01 941d877
fix: ord_explore allows unifrac family distances when phy_tree available
