### 2021-05-01 0cd95a5
0.7.4 and vignette updates
