+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.3.9016
+Version: 0.7.4
 Authors@R: 
     person(given = "David",
            family = "Barnett",
