+++ R/ord_explore.R
@@ -57,7 +57,6 @@
 #'     tax_fix() %>%
 #'     ord_explore()
 #'
-#'
 #'   # simple example with precalculated ordination #
 #'   data("enterotype")
 #'   taxa_names(enterotype)[1] <- "unclassified" # replaces the "-1" taxon name
@@ -84,13 +83,15 @@
 #'     tax_transform("clr") %>%
 #'     ord_calc(constraints = c("weight", "female"))
 #'
+#'   # label style arguments can be passed to ord_explore
 #'   constrained_aitchison_rda %>%
 #'     ord_explore(
-#'       plot_taxa = 1:5, tax_lab_style = list(size = 3),
-#'       constraint_lab_style = list(size = 4), auto_caption = 7
+#'       tax_lab_style = list(size = 3),
+#'       constraint_lab_style = list(size = 4), auto_caption = 6
 #'     )
 #'   # Try changing the point colour to bmi_group or similar
-#'   # Style interactively! (setting e.g. colour as an argument doesn't work)
+#'   # Style points interactively!
+#'   # (setting colour/shape/etc as arguments doesn't work)
 #'
 #'   # Another dataset, where "size" variable drives gradient on PC1
 #'   # Try setting size and/or alpha to correspond to "size"!
@@ -151,7 +152,10 @@ ord_explore <- function(data,
             .shiny-split-layout > div {
               overflow: visible;
             }
-            .selectize-input, .irs, .form-control {
+            .form-group {
+              margin-bottom: 10px;
+            }
+            .selectize-input, .selectize-dropdown, .irs, .form-control {
               font-size: 12px;
             }
             "
@@ -200,17 +204,6 @@ ord_explore <- function(data,
                 selected = c(sample_id, "SAMPLE")[[1]] # 'SAMPLE' if id = NULL
               )
             ),
-            shiny::splitLayout(
-              cellWidths = c("30%", "65%"),
-              shiny::helpText("Colour:"),
-              shiny::selectInput(
-                inputId = "ord_colour", label = NULL, selected = "azure4",
-                choices = list(
-                  Variable = init$vars$all,
-                  Fixed = grDevices::colors(distinct = TRUE)
-                )
-              )
-            ),
             # shape
             shiny::splitLayout(
               cellWidths = c("30%", "65%"),
@@ -220,6 +213,18 @@ ord_explore <- function(data,
                 choices = list(
                   Variable = init$vars$all,
                   Fixed = ggplot2_shapes()
+                )
+              )
+            ),
+            #### colour -------------------------------------------------------
+            shiny::splitLayout(
+              cellWidths = c("30%", "65%"),
+              shiny::helpText("Colour:"),
+              shiny::selectInput(
+                inputId = "ord_colour", label = NULL, selected = "azure4",
+                choices = list(
+                  Variable = init$vars$all,
+                  Fixed = grDevices::colors(distinct = TRUE)
                 )
               )
             ),
@@ -274,6 +279,29 @@ ord_explore <- function(data,
                   options = list(placeholder = "numeric var?")
                 )
               )
+            ),
+            #### additions ----------------------------------------------------
+            shiny::splitLayout(
+              cellWidths = c("20%", "75%"),
+              shiny::helpText("Add:"),
+              shiny::selectInput(
+                inputId = "add", label = NULL, selected = "nothing",
+                choices = c(
+                  "nothing", "ellipses (coloured)" = "ellipses",
+                  "taxa (PCA/RDA/CCA)" = "taxa"
+                )
[…]
