### 2021-05-03 cb01ebc
algolia search init
