+++ _pkgdown.yml
@@ -1,3 +1,11 @@
+url: https://david-barnett.github.io/microViz/
+template:
+  params:
+    docsearch:
+      api_key: '22e6f0a75b7b1ffae9bf5bc5b58864d7'
+      index_name: 'microViz'
+
+
 reference:
 
 - title: "Manipulating sample_data"
