+++ pkgdown.yml
@@ -8,5 +8,8 @@ articles:
   articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
   ordination: ordination.html
-last_built: 2021-05-03T13:55Z
+last_built: 2021-05-03T14:15Z
+urls:
+  reference: https://david-barnett.github.io/microViz//reference
+  article: https://david-barnett.github.io/microViz//articles
 
