+++ README.md
@@ -14,30 +14,35 @@ Status](https://img.shields.io/docker/cloud/build/barnettdavid/microviz-rocker-v
 
 <!-- badges: end -->
 
-:package: microViz is an R package for analysis and visualization of
-microbiome sequencing data. microViz functions are intended to be easy
-to use and flexible. microViz extends and complements popular microbial
-ecology packages like phyloseq, vegan, and microbiome.
+## Overview
 
-**See the documentation website for full details and examples:**
-<https://david-barnett.github.io/microViz/>
+:package: `microViz` is an R package for analysis and visualization of
+microbiome sequencing data.
 
--   The [“Get
+:hammer: `microViz` functions are intended to be easy to use and
+flexible.
+
+:microbe: `microViz` extends and complements popular microbial ecology
+packages like `phyloseq`, `vegan`, and `microbiome`.
+
+## Learn more
+
+:paperclip: This website is the best place for documentation and
+examples: <https://david-barnett.github.io/microViz/>
+
+-   **The [“Get
     started”](https://david-barnett.github.io/microViz/articles/articles/microViz.html)
-    page, and also [this
-    ReadMe](https://david-barnett.github.io/microViz/index.html), show a
-    few example analyses.
+    page** shows a few example analyses (as does this ReadMe)
 
--   The
-    [reference](https://david-barnett.github.io/microViz/reference/index.html)
-    page lists all functions and links to their documentation and
-    examples
+-   **The
+    [Reference](https://david-barnett.github.io/microViz/reference/index.html)
+    page** lists all functions and links to help pages and examples
 
--   Example video of [Interactive Shiny
-    app](https://david-barnett.github.io/microViz/articles/articles/Interactive-Ordination.html)
-    for exploring your ordination plots!
+-   **The
+    [Changelog](https://david-barnett.github.io/microViz/news/index.html)**
+    describes important changes in new microViz package versions
 
--   Other articles pages give tutorials and more complicated examples
+-   **The Articles pages** give tutorials and further examples
 
     -   [Working with phyloseq
         objects](https://david-barnett.github.io/microViz/articles/articles/phyloseq.html)
@@ -49,16 +54,15 @@ ecology packages like phyloseq, vegan, and microbiome.
         plots](https://david-barnett.github.io/microViz/articles/ordination.html)
         (e.g. PCA or PCoA)
 
-    -   [Visualising taxonomic compositions with customised
-        barplots](https://david-barnett.github.io/microViz/articles/articles/Visualising-compositions.html)
+    -   [Interactive ordination plots with
+        ord\_explore](https://david-barnett.github.io/microViz/articles/articles/Interactive-Ordination.html)
 
-    -   (more coming soon! Post on [GitHub
+    -   [Visualising taxonomic compositions with
+        comp\_barplot](https://david-barnett.github.io/microViz/articles/articles/Visualising-compositions.html)
+
+    -   More coming soon! Post on [GitHub
         discussions](https://github.com/david-barnett/microViz/discussions)
-        if you have questions/requests)
-
--   The
-    [changelog](https://david-barnett.github.io/microViz/news/index.html)
-    describes important changes in new microViz package versions
+        if you have questions/requests
 
 ## Installation
 
@@ -66,26 +70,30 @@ You can install the latest available microViz package version using the
 following instructions.
 
 ``` r
-# Windows users will need to have RTools installed so that your computer can build this package
-# Follow instructions here: http://jtleek.com/modules/01_DataScientistToolbox/02_10_rtools/
-
[…]
