+++ README.Rmd
@@ -25,19 +25,25 @@ options(width = 100)
 
 <!-- badges: end -->
 
-:package: microViz is an R package for analysis and visualization of microbiome sequencing data.
-microViz functions are intended to be easy to use and flexible.
-microViz extends and complements popular microbial ecology packages like phyloseq, vegan, and microbiome.
+## Overview
 
-**See the documentation website for full details and examples:** <https://david-barnett.github.io/microViz/>
+:package: `microViz` is an R package for analysis and visualization of microbiome sequencing data.
 
--   The ["Get started"](https://david-barnett.github.io/microViz/articles/articles/microViz.html) page, and also [this ReadMe](https://david-barnett.github.io/microViz/index.html), show a few example analyses.
+:hammer: `microViz` functions are intended to be easy to use and flexible.
 
--   The [reference](https://david-barnett.github.io/microViz/reference/index.html) page lists all functions and links to their documentation and examples
+:microbe: `microViz` extends and complements popular microbial ecology packages like `phyloseq`, `vegan`, and `microbiome`.
 
--   Example video of [Interactive Shiny app](https://david-barnett.github.io/microViz/articles/articles/Interactive-Ordination.html) for exploring your ordination plots!
+## Learn more
 
--   Other articles pages give tutorials and more complicated examples
+:paperclip: This website is the best place for documentation and examples: <https://david-barnett.github.io/microViz/>
+
+-   **The ["Get started"](https://david-barnett.github.io/microViz/articles/articles/microViz.html) page** shows a few example analyses (as does this ReadMe)
+
+-   **The [Reference](https://david-barnett.github.io/microViz/reference/index.html) page** lists all functions and links to help pages and examples
+
+-   **The [Changelog](https://david-barnett.github.io/microViz/news/index.html)** describes important changes in new microViz package versions
+
+-   **The Articles pages** give tutorials and further examples
 
     -   [Working with phyloseq objects](https://david-barnett.github.io/microViz/articles/articles/phyloseq.html)
 
@@ -45,37 +51,37 @@ microViz extends and complements popular microbial ecology packages like phylose
 
     -   [Creating ordination plots](https://david-barnett.github.io/microViz/articles/ordination.html) (e.g. PCA or PCoA)
 
-    -   [Visualising taxonomic compositions with customised barplots](https://david-barnett.github.io/microViz/articles/articles/Visualising-compositions.html)
+    -   [Interactive ordination plots with ord_explore](https://david-barnett.github.io/microViz/articles/articles/Interactive-Ordination.html)
 
-    -   (more coming soon! Post on [GitHub discussions](https://github.com/david-barnett/microViz/discussions) if you have questions/requests)
+    -   [Visualising taxonomic compositions with comp_barplot](https://david-barnett.github.io/microViz/articles/articles/Visualising-compositions.html)
 
--   The [changelog](https://david-barnett.github.io/microViz/news/index.html) describes important changes in new microViz package versions
+    -   More coming soon!
+        Post on [GitHub discussions](https://github.com/david-barnett/microViz/discussions) if you have questions/requests
 
 ## Installation
 
 You can install the latest available microViz package version using the following instructions.
 
 ``` r
-# Windows users will need to have RTools installed so that your computer can build this package
-# Follow instructions here: http://jtleek.com/modules/01_DataScientistToolbox/02_10_rtools/
-
-# macOS users might need to install xquartz to make the heatmaps work (ComplexHeatmaps package)
-# You can do this with homebrew, running the following command in your mac's Terminal: brew install --cask xquartz 
-
-# for installing from github you'll need the devtools package
+# Installing from github requires the devtools package
[…]
