+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
   ordination: ordination.html
-last_built: 2021-05-03T14:15Z
+last_built: 2021-05-06T07:16Z
 urls:
   reference: https://david-barnett.github.io/microViz//reference
   article: https://david-barnett.github.io/microViz//articles
