### 2021-05-06 c9d900c
clean ord_explore code
