+++ R/ord_explore.R
@@ -185,8 +185,7 @@ ord_explore <- function(data,
             ),
             shiny::h4("Ordination options"),
             shiny::splitLayout(
-              cellWidths = c("30%", "30%", "30%"),
-              shiny::helpText("Dims:"),
+              cellWidths = c("30%", "30%", "30%"), shiny::helpText("Dims:"),
               shiny::numericInput(
                 inputId = "x1", label = NULL,
                 value = 1, min = 1, max = 150, step = 1
@@ -197,8 +196,7 @@ ord_explore <- function(data,
               )
             ),
             shiny::splitLayout(
-              cellWidths = c("30%", "65%"),
-              shiny::helpText("Select:"),
+              cellWidths = c("30%", "65%"), shiny::helpText("Select:"),
               shiny::selectInput(
                 inputId = "id_var", label = NULL, choices = init$vars$all,
                 selected = c(sample_id, "SAMPLE")[[1]] # 'SAMPLE' if id = NULL
@@ -206,20 +204,17 @@ ord_explore <- function(data,
             ),
             # shape
             shiny::splitLayout(
-              cellWidths = c("30%", "65%"),
-              shiny::helpText("Shape:"),
+              cellWidths = c("30%", "65%"), shiny::helpText("Shape:"),
               shiny::selectInput(
                 inputId = "ord_shape", label = NULL, selected = "circle",
                 choices = list(
-                  Variable = init$vars$all,
-                  Fixed = ggplot2_shapes()
+                  Variable = init$vars$all, Fixed = ggplot2_shapes()
                 )
               )
             ),
             #### colour -------------------------------------------------------
             shiny::splitLayout(
-              cellWidths = c("30%", "65%"),
-              shiny::helpText("Colour:"),
+              cellWidths = c("30%", "65%"), shiny::helpText("Colour:"),
               shiny::selectInput(
                 inputId = "ord_colour", label = NULL, selected = "azure4",
                 choices = list(
@@ -245,8 +240,7 @@ ord_explore <- function(data,
             shiny::conditionalPanel(
               condition = "input.alphaFixed == false",
               shiny::splitLayout(
-                cellWidths = c("30%", "65%"),
-                shiny::helpText("Alpha:"),
+                cellWidths = c("30%", "65%"), shiny::helpText("Alpha:"),
                 shiny::selectizeInput(
                   inputId = "ord_alpha_var", label = NULL,
                   choices = init$vars$num, selected = NULL,
@@ -282,12 +276,12 @@ ord_explore <- function(data,
             ),
             #### additions ----------------------------------------------------
             shiny::splitLayout(
-              cellWidths = c("20%", "75%"),
-              shiny::helpText("Add:"),
+              cellWidths = c("20%", "75%"), shiny::helpText("Add:"),
               shiny::selectInput(
                 inputId = "add", label = NULL, selected = "nothing",
                 choices = c(
-                  "nothing", "ellipses (coloured)" = "ellipses",
+                  "nothing",
+                  "ellipses (coloured)" = "ellipses",
                   "taxa (PCA/RDA/CCA)" = "taxa"
                 )
               )
@@ -295,8 +289,7 @@ ord_explore <- function(data,
             shiny::conditionalPanel(
               condition = "input.add == 'taxa'",
               shiny::splitLayout(
-                cellWidths = c("50%", "45%"),
-                shiny::helpText("N labels:"),
+                cellWidths = c("50%", "45%"), shiny::helpText("N labels:"),
                 shiny::numericInput(
                   inputId = "nLabels", label = NULL, value = 3,
                   min = 1, max = 25, step = 1
[…]
