### 2021-05-06 c215b55
add tests for ord_explore helpers
