+++ tests/testthat/_snaps/ord_explore.md
@@ -1,0 +1,614 @@
+# ord_explore_init stays the same
+
+    Code
+      ord_explore_init(dietswap)
+    Output
+      $data
+      ps_extra object - a list with phyloseq and extras:
+      
+      phyloseq-class experiment-level object
+      otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
+      sample_data() Sample Data:       [ 222 samples by 9 sample variables ]
+      tax_table()   Taxonomy Table:    [ 130 taxa by 4 taxonomic ranks ]
+      
+      ps_extra info:
+      tax_agg = unique tax_transform = identity
+      
+      $info
+      $info$rank
+      [1] "unique"
+      
+      $info$trans
+      [1] "identity"
+      
+      $info$scale
+      [1] NA
+      
+      $info$dist
+      [1] "none"
+      
+      $info$ord
+      [1] "auto"
+      
+      $info$constraints
+      NULL
+      
+      $info$conditions
+      NULL
+      
+      $info$isCon
+      [1] FALSE
+      
+      
+      $vars
+      $vars$all
+      [1] "subject"                "sex"                    "nationality"           
+      [4] "group"                  "sample"                 "timepoint"             
+      [7] "timepoint.within.group" "bmi_group"              "SAMPLE"                
+      
+      $vars$num
+      [1] "timepoint"              "timepoint.within.group"
+      
+      $vars$cat
+      [1] "subject"     "sex"         "nationality" "group"       "sample"     
+      [6] "bmi_group"   "SAMPLE"     
+      
+      
+      $ranks
+      [1] "Phylum" "Family" "Genus"  "unique"
+      
+      $warn
+      [1] FALSE
+      
+
+---
+
+    Code
+      ord_explore_init(ord)
+    Output
+      $data
+      ps_extra object - a list with phyloseq and extras:
+      
+      phyloseq-class experiment-level object
+      otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
+      sample_data() Sample Data:       [ 222 samples by 11 sample variables ]
+      tax_table()   Taxonomy Table:    [ 130 taxa by 4 taxonomic ranks ]
+      
+      ps_extra info:
+      tax_agg = Genus tax_transform = clr
+      
+      ordination of class: rda cca 
+      rda(formula = OTU ~ weight + female, data = data)
+      constraints: weight+female
+      
+      $counts OTU Table: [ 130 taxa and 222 samples ]
+      
+      $info
+      $info$rank
+      [1] "Genus"
+      
+      $info$trans
+      [1] "clr"
+      
+      $info$scale
+      [1] NA
+      
+      $info$dist
+      [1] "none"
+      
+      $info$ord
+      [1] "RDA"
+      
+      $info$constraints
+      [1] "weight" "female"
+      
+      $info$conditions
+      NULL
+      
+      $info$isCon
+      [1] TRUE
+      
+      
+      $vars
+      $vars$all
+       [1] "subject"                "sex"                    "nationality"           
+       [4] "group"                  "sample"                 "timepoint"             
+       [7] "timepoint.within.group" "bmi_group"              "weight"                
+      [10] "female"                 "SAMPLE"                
+      
+      $vars$num
+      [1] "timepoint"              "timepoint.within.group" "weight"                
+      [4] "female"                
+      
+      $vars$cat
+      [1] "subject"     "sex"         "nationality" "group"       "sample"     
+      [6] "bmi_group"   "SAMPLE"     
+      
+      
+      $ranks
+      [1] "Phylum" "Family" "Genus"  "unique"
+      
+      $warn
[…]
