+++ README.Rmd
@@ -31,7 +31,7 @@ options(width = 100)
 
 :hammer: `microViz` functions are intended to be easy to use and flexible.
 
-:microbe: `microViz` extends and complements popular microbial ecology packages like `phyloseq`, `vegan`, and `microbiome`.
+:microscope: `microViz` extends and complements popular microbial ecology packages like `phyloseq`, `vegan`, and `microbiome`.
 
 ## Learn more
 
@@ -283,6 +283,6 @@ This project is released with a [Contributor Code of Conduct](https://david-barn
 
 ## Session info
 
-```{r session info}
-devtools::session_info()
+```{r session}
+sessionInfo()
 ```
