+++ README.Rmd
@@ -31,7 +31,7 @@ options(width = 100)
 
 :hammer: `microViz` functions are intended to be easy to use and flexible.
 
-:microscope: `microViz` extends and complements popular microbial ecology packages like `phyloseq`, `vegan`, and `microbiome`.
+:microscope: `microViz` extends and complements popular microbial ecology packages like `phyloseq`, `vegan`, & `microbiome`.
 
 ## Learn more
 
