+++ tests/testthat/apps/tax_fix_interactive/tests/shinytest/mytest.R
@@ -2,16 +2,21 @@ app <- ShinyDriver$new("../../", seed = 1)
 app$snapshotInit("mytest")
 
 app$setInputs(min_char = 0)
-tmp <- app$waitForValue("view_rows_all", iotype = "input", ignore = list(NULL))
+tmp <- app$waitForValue("view_rows_all", iotype = "input")
+tmp <- app$waitForShiny()
 app$snapshot()
 app$setInputs(selected = "Anaerostipes caccae et rel.")
+tmp <- app$waitForShiny()
 app$snapshot()
 app$setInputs(highlight = "Actinomycetaceae")
+tmp <- app$waitForShiny()
 app$snapshot()
 app$setInputs(tab = "fixed")
-tmp <- app$waitForValue("in_tt_rows_all", iotype = "input", ignore = list(NULL))
+tmp <- app$waitForValue("in_tt_rows_all", iotype = "input")
+tmp <- app$waitForShiny()
 app$snapshot()
 app$setInputs(tab = "tips")
+tmp <- app$waitForShiny()
 app$snapshot()
 rm(tmp)
 
