+++ R/tax_agg.R
@@ -55,6 +55,7 @@
 #' @export
 #'
 #' @seealso \code{\link{tax_fix}}
+#' @seealso \code{\link{tax_fix_interactive}}
 #' @seealso \code{\link{tax_transform}}
 #'
 #' @rdname tax_agg
@@ -70,14 +71,12 @@
 #'   ps_get() %>%
 #'   tax_table()
 #'
-#' # this won't aggregate taxa, but just adds a new rank called unique, equal to taxa_names
-#' tax_agg(ps = dietswap, rank = NA, add_unique = TRUE)
-#' identical(tax_agg(dietswap, NA, add_unique = TRUE), tax_agg(dietswap, "unique")) # TRUE
-#'
 #' # create some missing values
 #' tax_table(dietswap)[3:7, "Genus"] <- "g__"
-#' # this will produce an error, instructing the user to use fill_unknowns
+#'
+#' # this will produce an error, instructing the user to use tax_fix
 #' # tax_agg(ps = dietswap, "Genus")
+#'
 #' # this will then work:
 #' dietswap %>%
 #'   tax_fix() %>%
@@ -91,7 +90,14 @@
 #'   tax_fix(unknowns = "some_unknown_family") %>%
 #'   tax_agg("Family")
 #'
-#' # try tax_fix_interactive() for more help
+#' # try tax_fix_interactive() to help you find and fix all the uninformative
+#' # and converging values in your taxonomy table.
+#'
+#' # the code below won't aggregate taxa,
+#' but just adds a new rank called unique, equal to taxa_names
+#' tax_agg(ps = dietswap, rank = NA, add_unique = TRUE)
+#' identical(tax_agg(dietswap, NA, add_unique = TRUE), tax_agg(dietswap, "unique")) # TRUE
+#'
 tax_agg <- function(ps,
                     rank = NA,
                     sort_by = NA,
@@ -108,17 +114,6 @@ tax_agg <- function(ps,
   # store taxa orientation info for restoration to original setup before return
   taxa_were_rows <- phyloseq::taxa_are_rows(ps)
 
-  # set up messages sent by multiple stop calls
-  unknowns_message <-
-    "\nTo fix, try tax_fix() or tax_fix_interactive() before this function."
-
-  ranks <- phyloseq::rank_names(ps)
-  bad_rank_message <-
-    paste0(
-      "\nrank should be NA, 'unique' or one of:\n",
-      paste(ranks, collapse = " / ")
-    )
-
   # only do most things if rank is not NA/unique
   if (identical(rank, NA)) rank <- "unique"
   if (identical(rank, "unique")) {
@@ -130,76 +125,56 @@ tax_agg <- function(ps,
     ps@phy_tree <- NULL
     ps@refseq <- NULL
 
-    # check valid rank and get rank index
-    if (!identical(length(rank), 1L) && !inherits(rank, "character")) {
-      stop(bad_rank_message)
-    }
-    rank_index <- base::match(x = rank, table = ranks, nomatch = 0L)
-    if (identical(rank_index, 0L)) {
-      stop("rank is: ", rank, bad_rank_message)
-    }
+    ranks <- phyloseq::rank_names(ps)
+    # get rank index from vector of available ranks
+    rank_index <- getRankIndex(rank = rank, ranks = ranks)
 
-    # remove lower tax_table ranks
-    tt <- unclass(phyloseq::tax_table(ps))
-    tt <- tt[, seq_len(rank_index), drop = FALSE]
-    tt_df <- as.data.frame.matrix(tt, optional = TRUE, make.names = FALSE)
-    tt_df[[".taxID."]] <- tt_df[, rank_index]
-
-    # check NAs in chosen rank
[…]
