+++ vignettes/articles/tax-fixing.Rmd
@@ -9,8 +9,6 @@ knitr::opts_chunk$set(
 )
 ```
 
-------------------------------------------------------------------------
-
 ## Quick start / TLDR
 
 Use `tax_fix()` on your phyloseq data with default arguments to repair most `tax_table` problems (missing or uninformative values). If you still encounter errors using e.g. `tax_agg`, try using the Shiny app `tax_fix_interactive()` to help you generate `tax_fix` code that will fix your particular `tax_table` problems.
@@ -38,16 +36,16 @@ pseq
 The Species rank appears to be blank for many entries. This is a problem you may well encounter in your data: unique sequences or OTUs often cannot be annotated at lower taxonomic ranks.
 
 ```{r example tt look}
-tax_table(pseq)[1:10, 2:7]
+tax_table(pseq)[40:54, 4:7] # highest 3 ranks not shown, to save space
 ```
 
-If we would try to aggregate at Genus or Family rank level, we discover that blank values at these ranks prevent taxonomic aggregation.
+If we would try to aggregate at Genus or Family rank level, we discover that blank values at these ranks prevent taxonomic aggregation. This is because, for example, it looks like OTU.43 and OTU.54 share the same (empty) Genus name, "", despite being different at a higher rank, Family.
 
 ```{r tax_agg fail}
 # tax_agg(pseq, rank = "Family") # this fails, and sends (helpful) messages about taxa problems
 ```
 
-So we should run `tax_fix` first, which will fix most problems (with default settings), allowing taxa to be aggregated successfully (at any rank).
+So we should run `tax_fix` first, which will fix **most** problems with default settings, allowing taxa to be aggregated successfully (at any rank). If you still have errors when using tax_agg after tax_fix, carefully read the error and accompanying messages. Often you can copy suggested tax_fix code from the tax_agg error. You should generally also have a look around your tax_table for other uninformative values, using [tax_fix_interactive](https://david-barnett.github.io/microViz/articles/articles/tax-fixing.html#interactive-solutions).
 
 ```{r tax_agg success}
 pseq %>% tax_fix() %>% tax_agg(rank = "Family")
@@ -55,7 +53,7 @@ pseq %>% tax_fix() %>% tax_agg(rank = "Family")
 
 ### What does tax_fix do?
 
-`tax_fix` finds searches all the ranks of the phyloseq object `tax_table` for:
+`tax_fix` searches all the ranks of the phyloseq object `tax_table` for:
 
 -   short values, like "g\_\_", "", " ", etc. (any with fewer characters than min_length)
 
@@ -63,14 +61,14 @@ pseq %>% tax_fix() %>% tax_agg(rank = "Family")
 
 -   NAs
 
-`tax_fix` replaces these values with the next higher taxonomic rank, e.g. an "unknown" Family within the Order Clostridiales will be renamed "Clostridiales Order".
+`tax_fix` replaces these values with the next higher taxonomic rank, e.g. an "unknown" Family within the Order Clostridiales will be renamed "Clostridiales Order", as seen below.
 
 ```{r tax_fix look}
 pseq %>% tax_fix(min_length = 4) %>% 
   tax_agg("Family") %>% 
   ps_get() %>% # gets the phyloseq from the ps_extra list
   tax_table() %>% 
-  .[1:8, 2:5] # removes the Kingdom rank and shows only first 8 rows for nice printing
+  .[1:8, 3:5] # removes the first 2 ranks and shows only first 8 rows for nice printing
 ```
 
 ### Interactive solutions
@@ -85,13 +83,16 @@ tax_fix_interactive(example_data)
 
 ### Other possible problems
 
[…]
