+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
   ordination: ordination.html
-last_built: 2021-05-06T11:57Z
+last_built: 2021-05-07T20:36Z
 urls:
   reference: https://david-barnett.github.io/microViz//reference
   article: https://david-barnett.github.io/microViz//articles
