+++ README.md
@@ -71,15 +71,14 @@ following instructions.
 
 ``` r
 # Installing from github requires the devtools package
-install.packages("devtools", type = "binary") # (not binary if you're on linux)
+install.packages("devtools") 
 
 # You can install the latest versions of phyloseq and microbiome from Bioconductor:
-if (!requireNamespace("BiocManager", quietly = TRUE))
- install.packages("BiocManager", type = "binary") # (not binary if you're on linux)
+if (!requireNamespace("BiocManager", quietly = TRUE)) install.packages("BiocManager") 
 BiocManager::install(c("phyloseq", "microbiome"))
 
 # To install the latest "released" version of this package
-devtools::install_github("david-barnett/microViz@0.7.4") # check 0.7.4 is the latest release
+devtools::install_github("david-barnett/microViz@0.7.5") # check 0.7.5 is the latest release
 
 # To install the very latest version:
 devtools::install_github("david-barnett/microViz")
@@ -132,6 +131,7 @@ ord_explore(pseq) # gif generated with microViz version 0.7.4 (plays at 1.75x sp
 ``` r
 library(phyloseq)
 library(dplyr)
+#> Warning: package 'dplyr' was built under R version 4.0.4
 #> 
 #> Attaching package: 'dplyr'
 #> The following objects are masked from 'package:stats':
@@ -190,7 +190,7 @@ Maybe visually inspecting all your samples isn’t quite what you want.
 Ordination methods can also help you to visualise if overall microbial
 ecosystem composition differs markedly between groups, e.g. BMI.
 
-Here is one option to try first:
+Here is one option as an example:
 
 1.  Filter out rare taxa (e.g. remove Genera not present in at least 10%
     of samples) - use `tax_filter()`
@@ -265,8 +265,8 @@ aitchison_perm <- dist_permanova(
   variables = "bmi_group + female"
 )
 #> Dropping samples with missings: 2
-#> 2021-05-06 13:53:45 - Starting PERMANOVA with 99 perms with 1 processes
-#> 2021-05-06 13:53:45 - Finished PERMANOVA
+#> 2021-05-09 13:09:06 - Starting PERMANOVA with 99 perms with 1 processes
+#> 2021-05-09 13:09:06 - Finished PERMANOVA
 # view the permanova results
 perm_get(aitchison_perm) %>% as.data.frame()
 #>            Df   SumOfSqs         R2        F Pr(>F)
@@ -294,8 +294,8 @@ your permanova directly using the ord\_plot function with constraints.
 ``` r
 perm2 <- dist_permanova(data = aitchison_dists, variables = c("weight", "female"), seed = 321)
 #> Dropping samples with missings: 2
-#> 2021-05-06 13:53:45 - Starting PERMANOVA with 999 perms with 1 processes
-#> 2021-05-06 13:53:46 - Finished PERMANOVA
+#> 2021-05-09 13:09:06 - Starting PERMANOVA with 999 perms with 1 processes
+#> 2021-05-09 13:09:07 - Finished PERMANOVA
 perm_get(perm2)
 #> Permutation test for adonis under reduced model
 #> Marginal effects of terms
@@ -398,21 +398,21 @@ sessionInfo()
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
 #> 
 #> other attached packages:
[…]
