+++ README.Rmd
@@ -64,15 +64,14 @@ You can install the latest available microViz package version using the followin
 
 ``` r
 # Installing from github requires the devtools package
-install.packages("devtools", type = "binary") # (not binary if you're on linux)
+install.packages("devtools") 
 
 # You can install the latest versions of phyloseq and microbiome from Bioconductor:
-if (!requireNamespace("BiocManager", quietly = TRUE))
- install.packages("BiocManager", type = "binary") # (not binary if you're on linux)
+if (!requireNamespace("BiocManager", quietly = TRUE)) install.packages("BiocManager") 
 BiocManager::install(c("phyloseq", "microbiome"))
 
 # To install the latest "released" version of this package
-devtools::install_github("david-barnett/microViz@0.7.4") # check 0.7.4 is the latest release
+devtools::install_github("david-barnett/microViz@0.7.5") # check 0.7.5 is the latest release
 
 # To install the very latest version:
 devtools::install_github("david-barnett/microViz")
@@ -158,7 +157,7 @@ Maybe visually inspecting all your samples isn't quite what you want.
 Ordination methods can also help you to visualise if overall microbial ecosystem composition differs markedly between groups, e.g.
 BMI.
 
-Here is one option to try first:
+Here is one option as an example:
 
 1.  Filter out rare taxa (e.g. remove Genera not present in at least 10% of samples) - use `tax_filter()`
 2.  Aggregate the taxa into bacterial families (for example) - use `tax_agg()`
