### 2021-05-09 b414ad5
tests and doc typos
