### 2021-05-09 #3201 @david-barnett
Hi all,

Firstly, many thanks to marypiper and yoonjeongcha for agreeing to review my work :slightly_smiling_face:

Secondly, over the last couple of weeks I have made some, mostly small, changes to microViz in response to user requests and reports, which are listed in the [changelog](https://david-barnett.github.io/microViz/news/index.html)

Should I / can I somehow update the version in Whedon's comment to the current version? 0.7.5
