### 2021-05-11 6e6667b
fix: cor heatmap df_to_numeric_matrix helper error message itself generated an error!
