+++ NEWS.md
@@ -1,3 +1,5 @@
+# microViz (development version)
+
 # microViz 0.7.5
 
 - `ord_explore` can now draw stat_ellipse or taxa loading vectors
