+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.5.9000
+Version: 0.7.5.9001
 Authors@R: 
     person(given = "David",
            family = "Barnett",
