### 2021-05-15 4bcc7e9
ord_explore refactoring: extract barplot helpers and make names clearer
