+++ R/ord_explore.R
@@ -412,9 +412,9 @@ ord_explore <- function(data,
         shiny::hr(),
         shiny::renderPrint({
           ord_code(
-            rank = ordSelected$rank, trans = ordSelected$trans,
-            dist = ordSelected$dist, ord = ordSelected$ord,
-            const = ordSelected$const, conds = ordSelected$conds,
+            rank = ord1chosen$rank, trans = ord1chosen$trans,
+            dist = ord1chosen$dist, ord = ord1chosen$ord,
+            const = ord1chosen$const, conds = ord1chosen$conds,
             x = input$x1, y = input$y1, colour = input$ord_colour,
             fill = input$ord_colour, # TODO make fill configurable
             shape = input$ord_shape, alpha = alpha(), size = size(),
@@ -443,36 +443,36 @@ ord_explore <- function(data,
         shiny::hr(),
         shiny::selectizeInput(
           inputId = "rank", label = "Taxonomic Rank",
-          selected = ordSelected$rank, choices = ordChoices$rank
+          selected = ord1chosen$rank, choices = ordChoices$rank
         ),
         shiny::selectizeInput(
           inputId = "trans", label = "Taxa transformation",
-          choices = ordChoices$trans, selected = ordSelected$trans
+          choices = ordChoices$trans, selected = ord1chosen$trans
         ),
         shiny::selectizeInput(
           inputId = "dist", label = "Distance / Dissimilarity",
-          choices = ordChoices$dist, selected = ordSelected$dist
+          choices = ordChoices$dist, selected = ord1chosen$dist
         ),
         shiny::checkboxInput(
           inputId = "isCon", label = "Constrain or condition ordination?",
-          value = ordSelected$isCon
+          value = ord1chosen$isCon
         ),
         shiny::conditionalPanel(
           condition = "input.isCon == true",
           shiny::selectizeInput(
             inputId = "const", label = "Constraints", multiple = TRUE,
-            choices = ordChoices$const, selected = ordSelected$const,
+            choices = ordChoices$const, selected = ord1chosen$const,
             options = list(placeholder = "numeric vars?")
           ),
           shiny::selectizeInput(
             inputId = "conds", label = "Conditions", multiple = TRUE,
-            choices = ordChoices$conds, selected = ordSelected$conds,
+            choices = ordChoices$conds, selected = ord1chosen$conds,
             options = list(placeholder = "numeric vars?")
           )
         ),
         shiny::selectizeInput(
           inputId = "method", label = "Ordination method",
-          choices = ordChoices$ord, selected = ordSelected$ord
+          choices = ordChoices$ord, selected = ord1chosen$ord
         ),
         footer = shiny::tagList(
           shiny::modalButton("Cancel", icon = shiny::icon("times")),
@@ -509,7 +509,7 @@ ord_explore <- function(data,
     ### selected --------------------------------------------------------------
     # for remembering selected and possible choices in modal selectize inputs
     #### initialise selected choices ------------------------------------------
-    ordSelected <- shiny::reactiveValues(
+    ord1chosen <- shiny::reactiveValues(
       rank = init$info$rank, trans = init$info$trans,
       # scale = if (is.na(info$scale)) "neither" else info$scale,
       dist = init$info$dist, ord = init$info$ord,
@@ -522,15 +522,15 @@ ord_explore <- function(data,
     shiny::observeEvent(
       eventExpr = input$build,
       handlerExpr = {
-        ordSelected$rank <- input$rank
-        ordSelected$trans <- input$trans
[…]
