### 2021-05-15 1e6dc97
ord_explore barplot facet grid with space=free instead of wrap with fixed space
