### 2021-05-15 d0c091d
ord_explore cleaning: move other helpers to end
