### 2021-05-15 8a658a8
ord_explore plot css style improvements
