+++ NAMESPACE
@@ -41,6 +41,7 @@ export(ps_reorder)
 export(ps_select)
 export(ps_seriate)
 export(samdat_tbl)
+export(stat_chull)
 export(tax_agg)
 export(tax_anno)
 export(tax_filter)
