### 2021-05-17 5f80724
fix: ord_plot change class to inherit
