### 2021-05-17 45d403c
feat: add_paths function for ord_plot and ord_explore enhancement with time series data
