+++ NAMESPACE
@@ -3,6 +3,7 @@
 S3method(print,ps_extra)
 S3method(print,ps_extra_info)
 export("%>%")
+export(add_paths)
 export(anno_abund)
 export(anno_prev)
 export(anno_var_box)
