+++ R/stat_chull.R
@@ -17,18 +17,19 @@
 #' @export
 #' @examples
 #' library(ggplot2)
-#' corncob::ibd_phylo %>% tax_fix() %>%
+#' corncob::ibd_phylo %>%
+#'   tax_fix() %>%
 #'   tax_transform(rank = "Genus", transformation = "clr") %>%
 #'   ord_calc(method = "PCA") %>%
 #'   ord_plot(colour = "DiseaseState", shape = "DiseaseState", alpha = 0.5) +
 #'   stat_chull(aes(colour = DiseaseState))
 #'
-#' corncob::ibd_phylo %>% tax_fix() %>%
+#' corncob::ibd_phylo %>%
+#'   tax_fix() %>%
 #'   tax_transform(rank = "Genus", transformation = "clr") %>%
 #'   ord_calc(method = "PCA") %>%
 #'   ord_plot(colour = "DiseaseState", shape = "DiseaseState", alpha = 0.5) +
 #'   stat_chull(aes(colour = DiseaseState, fill = DiseaseState), alpha = 0.1)
-#'
 stat_chull <- function(mapping = NULL, data = NULL, geom = "polygonHollow",
                        position = "identity", na.rm = FALSE, show.legend = NA,
                        inherit.aes = TRUE, ...) {
