### 2021-05-17 0d29ffd
cleaning: move ord_plot ord and ps size comparison check earlier
