+++ man/stat_chull.Rd
@@ -72,18 +72,19 @@ This is a ggplot2 extension - slightly modified from the original code found her
 }
 \examples{
 library(ggplot2)
-corncob::ibd_phylo \%>\% tax_fix() \%>\%
+corncob::ibd_phylo \%>\%
+  tax_fix() \%>\%
   tax_transform(rank = "Genus", transformation = "clr") \%>\%
   ord_calc(method = "PCA") \%>\%
   ord_plot(colour = "DiseaseState", shape = "DiseaseState", alpha = 0.5) +
   stat_chull(aes(colour = DiseaseState))
 
-corncob::ibd_phylo \%>\% tax_fix() \%>\%
+corncob::ibd_phylo \%>\%
+  tax_fix() \%>\%
   tax_transform(rank = "Genus", transformation = "clr") \%>\%
   ord_calc(method = "PCA") \%>\%
   ord_plot(colour = "DiseaseState", shape = "DiseaseState", alpha = 0.5) +
   stat_chull(aes(colour = DiseaseState, fill = DiseaseState), alpha = 0.1)
-
 }
 \seealso{
 \code{ggplot2::\link{stat_ellipse}}
