### 2021-05-18 052db60
ord_explore ord_code now outputs chulls and paths if required
