+++ R/ord_explore.R
@@ -97,7 +97,10 @@
 #'   # subject. If we arrange by timepoint first (!!!), we can use the "paths"
 #'   # additional plot layer from the ord_explore "Add:" menu to track
 #'   # individual subjects over time.
-#'   dietswap %>% ps_arrange(timepoint) %>% tax_fix() %>% ord_explore()
+#'   dietswap %>%
+#'     ps_arrange(timepoint) %>%
+#'     tax_fix() %>%
+#'     ord_explore()
 #'
 #'
 #'   # Another dataset, where "size" variable drives gradient on PC1
@@ -444,7 +447,8 @@ ord_explore <- function(data,
             x = input$x1, y = input$y1, colour = input$ord_colour,
             fill = input$ord_colour, # TODO make fill configurable
             shape = input$ord_shape, alpha = alpha(), size = size(),
-            plot_taxa = plot_taxa(), ellipses = ellipses()
+            plot_taxa = plot_taxa(), ellipses = ellipses(),
+            chulls = chulls(), paths = paths()
           )
         }),
         shiny::hr(),
@@ -739,7 +743,7 @@ ord_explore <- function(data,
       input$add == "chulls" & input$ord_colour %in% init$vars$all
     })
     paths <- shiny::reactive({
-      if (input$add == "paths" && length(input$pathGroupsChosen) > 0){
+      if (input$add == "paths" && length(input$pathGroupsChosen) > 0) {
         list(
           id_var = shiny::isolate(input$pathGroupID),
           id_values = input$pathGroupsChosen,
@@ -1013,8 +1017,8 @@ ord_ggplot <- function(ord, x, y, shape, size, colour, alpha, id,
     if (chulls) p1 <- p1 + stat_chull(ggplot2::aes(colour = .data[[colour]]))
 
     # optionally add (time) paths to selected groups
-    if (!identical(paths, NULL)){
-      if (paths$colour %in% paths$all_vars){
+    if (!identical(paths, NULL)) {
+      if (paths$colour %in% paths$all_vars) {
         p1 <- add_paths(
           ggplot = p1, id_var = paths$id_var, id_values = paths$id_values,
           mapping = ggplot2::aes(colour = .data[[paths$colour]])
@@ -1031,7 +1035,7 @@ ord_ggplot <- function(ord, x, y, shape, size, colour, alpha, id,
 }
 
 # create girafe interactive plot from ggplot ord_plot
-ord_girafe <- function(gg, width, height){
+ord_girafe <- function(gg, width, height) {
   ggiraph::girafe(
     ggobj = gg, width_svg = width, height_svg = height,
     options = list(
@@ -1157,7 +1161,8 @@ trans_choices <- function(type) {
 ### code modal helpers --------------------------------------------------------
 # generate code-styled text for reproducing ordination plot
 ord_code <- function(rank, trans, dist, ord, const, conds, x, y,
-                     colour, fill, shape, alpha, size, plot_taxa, ellipses) {
+                     colour, fill, shape, alpha, size,
+                     plot_taxa, ellipses, chulls, paths) {
   # prepare dist_calc line if distance needed
   dist_calc_line <- ord_code_dist(dist)
 
@@ -1181,8 +1186,15 @@ ord_code <- function(rank, trans, dist, ord, const, conds, x, y,
   if (!is.numeric(alpha)) alpha <- paste0('"', alpha, '"')
   if (!is.numeric(size)) size <- paste0('"', size, '"')
 
-  # prepare extra stat_ellipse lines for end of code if necessary
[…]
