+++ tests/testthat/_snaps/ord_explore.md
@@ -329,7 +329,7 @@
     Code
       ord_code(rank = "Genus", trans = "identity", dist = "none", ord = "RDA", const = c,
         conds = NULL, x = 1, y = 2, colour = "v", fill = "v", shape = "var", alpha = a,
-        size = 1, plot_taxa = p, ellipses = FALSE)
+        size = 1, plot_taxa = p, ellipses = FALSE, chulls = FALSE, paths = NULL)
     Output
       your_phyloseq %>%
        tax_transform(rank = "Genus", transformation = "identity") %>%
@@ -352,7 +352,7 @@
     Code
       ord_code(rank = "Genus", trans = "identity", dist = "none", ord = "RDA", const = c,
         conds = NULL, x = 1, y = 2, colour = "v", fill = "v", shape = "var", alpha = a,
-        size = 1, plot_taxa = p, ellipses = FALSE)
+        size = 1, plot_taxa = p, ellipses = FALSE, chulls = FALSE, paths = NULL)
     Output
       your_phyloseq %>%
        tax_transform(rank = "Genus", transformation = "identity") %>%
@@ -376,7 +376,7 @@
     Code
       ord_code(rank = "Genus", trans = "identity", dist = "none", ord = "RDA", const = c,
         conds = NULL, x = 1, y = 2, colour = "v", fill = "v", shape = "var", alpha = a,
-        size = 1, plot_taxa = p, ellipses = FALSE)
+        size = 1, plot_taxa = p, ellipses = FALSE, chulls = FALSE, paths = NULL)
     Output
       your_phyloseq %>%
        tax_transform(rank = "Genus", transformation = "identity") %>%
@@ -399,7 +399,7 @@
     Code
       ord_code(rank = "Genus", trans = "identity", dist = "none", ord = "RDA", const = c,
         conds = NULL, x = 1, y = 2, colour = "v", fill = "v", shape = "var", alpha = a,
-        size = 1, plot_taxa = p, ellipses = FALSE)
+        size = 1, plot_taxa = p, ellipses = FALSE, chulls = FALSE, paths = NULL)
     Output
       your_phyloseq %>%
        tax_transform(rank = "Genus", transformation = "identity") %>%
@@ -423,7 +423,7 @@
     Code
       ord_code(rank = "Genus", trans = "identity", dist = "none", ord = "RDA", const = c,
         conds = NULL, x = 1, y = 2, colour = "v", fill = "v", shape = "var", alpha = a,
-        size = 1, plot_taxa = p, ellipses = FALSE)
+        size = 1, plot_taxa = p, ellipses = FALSE, chulls = FALSE, paths = NULL)
     Output
       your_phyloseq %>%
        tax_transform(rank = "Genus", transformation = "identity") %>%
@@ -447,7 +447,7 @@
     Code
       ord_code(rank = "Genus", trans = "identity", dist = "none", ord = "RDA", const = c,
         conds = NULL, x = 1, y = 2, colour = "v", fill = "v", shape = "var", alpha = a,
-        size = 1, plot_taxa = p, ellipses = FALSE)
+        size = 1, plot_taxa = p, ellipses = FALSE, chulls = FALSE, paths = NULL)
     Output
       your_phyloseq %>%
        tax_transform(rank = "Genus", transformation = "identity") %>%
@@ -472,7 +472,7 @@
     Code
       ord_code(rank = "Genus", trans = "identity", dist = "none", ord = "RDA", const = c,
         conds = NULL, x = 1, y = 2, colour = "v", fill = "v", shape = "var", alpha = a,
-        size = 1, plot_taxa = p, ellipses = FALSE)
[…]
