+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.5.9008
+Version: 0.7.5.9009
 Authors@R: 
     person(given = "David",
            family = "Barnett",
