### 2021-05-18 2368a02
better testing for ord_explore
