+++ tests/testthat/apps/ord_explore/tests/shinytest/smoketest.R
@@ -1,5 +1,0 @@
-app <- ShinyDriver$new("../../", seed = 1)
-app$snapshotInit("smoketest")
-
-app$setInputs(interactive = FALSE)
-app$snapshot(list(output = "comps_gg"))
