### 2021-05-22 david-barnett/microViz#25 @marypiper
RShiny app problems
When testing out the RShiny app functionality with the corncob dataset, few of the Selections worked with the app or didn't change the plot. For example, for 'shape', the race option (in addition to many others) generated the error: 'Error: ids don't have the same length than str (most often, it occurs because of clipping)'. Also, the lasso it doesn't always allow accurate selection - the data points I encircle are not the only points selected. Not sure why some of the time it works and other times it doesn't.
