### 2021-05-22 david-barnett/microViz#26 @marypiper
Statement of need 
Hi David,

Upon reading through your statement of need, I had some questions. I went by each statement to examine the novelty added by your tool, and maybe you can comment on these:

- Statement of need: microViz extends the set of distance measures (including e.g. Generalized UniFrac and Aitchison) of phyloseq and microbiome
    - Question: phyloseq has 44 supported distance methods  including UniFrac, what is provided that is not already included, and what additional uses do those distance measures provide?
- Statement of need: microViz makes use of further ordination calculation and visualization methods more accessible, such as (distance-based) redundancy analysis (RDA) and canonical correspondence analysis   (CCA)
     - Question: the microbiome R package has options for RDA and CCA , what additional functionality does microViz provide?
- Statement of need: microViz generates bi-plots and tri-plots with the ggplot2 R package
    - Question: the other standard microbiome packages generate bi-plots, what additional functionality does microViz provide?
