### 2021-05-22 #3201 @marypiper
Hi @david-barnett, just wanted to let you know that I have posted a couple of issues related to questions I had regarding the statement of need (https://github.com/david-barnett/microViz/issues/26) and the functionality of the RShiny app (https://github.com/david-barnett/microViz/issues/25). Thanks!
