+++ manuscript/microViz.bib
@@ -31,10 +31,11 @@
 	pages = {1--15}
 }
 
-@book{lahti_microbiome_2012,
-	title = {microbiome {R} package},
-	author = {Lahti, Leo and Shetty, Sudarshan},
-	year = {2012}
+@Misc{lahti_microbiome_2012,
+	title = {microbiome R package: tools for microbiome analysis in R},
+  author = {Lahti, Leo and Shetty, Sudarshan},
+  year = {2012-2019},
+  url = {https://github.com/microbiome/microbiome}
 }
 
 @article{mcmurdie_phyloseq_2013,
