+++ vignettes/ordination.Rmd
@@ -18,11 +18,13 @@ knitr::opts_chunk$set(
 
 This article will show you how to create and customise ordination plots, like [PCA](#pca) and [RDA](#rda), with microViz.
 
+For an even quicker start with ordinating your phyloseq data, check out the `ord_explore` [Shiny app](https://david-barnett.github.io/microViz/articles/articles/Interactive-Ordination.html), which allows you to create ordination plots with a point and click interface, and generates `ord_plot` code for you to copy. Return to this article to understand more about creating and customising your ordination plotting script.
+
 ```{r setup}
 library(phyloseq)
 library(ggplot2)
 library(microViz)
-knitr::opts_chunk$set(fig.width = 7, fig.height = 5)
+knitr::opts_chunk$set(fig.width = 7, fig.height = 6)
 ```
 
 We will use example data from stool samples from an inflammatory bowel disease (IBD) study, borrowed from the great `corncob` package. See the article about [working with phyloseq objects](https://david-barnett.github.io/microViz/articles/articles/phyloseq.html) if you want to get started with your own data, or just to learn more about manipulating phyloseq objects with microViz.
@@ -32,18 +34,18 @@ ibd <- corncob::ibd_phylo
 ibd
 ```
 
-First we check the phyloseq object and fix any uninformative tax_table entries (see the article on [fixing your tax_table](https://david-barnett.github.io/microViz/articles/articles/tax-fixing.html) for more info).
+First we fix any uninformative tax_table entries and check the phyloseq object (see the article on [fixing your tax_table](https://david-barnett.github.io/microViz/articles/articles/tax-fixing.html) for more info).
 
 ```{r fix it}
-ibd <- phyloseq_validate(ibd, remove_undetected = TRUE)
 ibd <- tax_fix(ibd) # try tax_fix_interactive if you have problems with your own data
+ibd <- phyloseq_validate(ibd, remove_undetected = TRUE)
 ```
 
 ## Motivation
 
 Ordination plots are a great way to see any clustering or other patterns of microbiota (dis)similarity in (many) samples. Ordinations like PCA or PCoA show the largest patterns of variation in your data, and constrained ordination techniques like RDA or CCA can show you microbial variation that could be explained by other variables in your sample_data (but interpret constrained ordinations with care, and ideally test for the statistical significance of any hypothesised associations using a method like PERMANOVA, see `dist_permanova()`).
 
-Check out the [Gusta Me website](https://sites.google.com/site/mb3gustame/) for a gentle theoretical introduction to [PCA](https://sites.google.com/site/mb3gustame/indirect-gradient-analysis/pca), [PCoA](https://sites.google.com/site/mb3gustame/dissimilarity-based-methods/principal-coordinates-analysis), [RDA](https://sites.google.com/site/mb3gustame/constrained-analyses/rda), [CCA](https://sites.google.com/site/mb3gustame/constrained-analyses/cca) and more.
[…]
