### 2021-05-28 0006c75
JOSS manuscript improvements
