+++ manuscript/paper.md
@@ -14,11 +14,11 @@ date: "7th April 2021"
 affiliations:
 - name: Maastricht Centre for Systems Biology (MaCSBio)
   index: 1
-- name: Maastricht University Medical Centre+, Department of Medical Microbiology
+- name: Maastricht University Medical Center+, Department of Medical Microbiology
   index: 2
-- name: Public Health Research Institute (CAPHRI)
+- name: Maastricht University, Public Health Research Institute (CAPHRI)
   index: 3
-- name: School of Nutrition and Translational Research in Metabolism (NUTRIM)
+- name: Maastricht University, School of Nutrition and Translational Research in Metabolism (NUTRIM)
   index: 4
 bibliography: microViz.bib
 tags:
@@ -27,45 +27,34 @@ tags:
 - visualization
 ---
 
-# Summary 
+# Summary
 
-microViz is an R package for the statistical analysis and visualization of microbiota data. This package extends the functionality of popular microbial ecosystem data analysis R packages, including phyloseq [@mcmurdie_phyloseq_2013], vegan [@oksanen_vegan_2020] and microbiome [@lahti_microbiome_2012]. 
-microViz provides a selection of powerful additions to the toolbox of researchers already familiar with phyloseq and microbiome, as well as assisting researchers with less R programming experience to independently explore and analyse their data and to generate publication-ready figures.
+microViz is an R package for the statistical analysis and visualization of microbiota data. This package extends the functionality of popular microbial ecosystem data analysis R packages, including phyloseq [@mcmurdie_phyloseq_2013], vegan [@oksanen_vegan_2020] and microbiome [@lahti_microbiome_2012]. microViz provides a selection of powerful additions to the toolbox of researchers already familiar with phyloseq and microbiome, as well as assisting researchers with less R programming experience to independently explore and analyse their data and to generate publication-ready figures.
 
 The tools offered by microViz include:
 
-- A Shiny app [@shiny] for interactive exploration of microbiota data within R, pairing ordination plots with abundance bar charts 
-- Easy to use functions for generating publication-ready ordination plots with ggplot2 [@wickham_ggplot2_2016], accommodating constrained and partial ordination, bi-plots and tri-plots, and automatic captioning designed to promote methodological transparency and reproducibility
-- A novel visualization approach pairing ordination plots with circular bar charts (iris plots) for comprehensive, intuitive and compact visualization of the similarity and composition of hundreds of microbial ecosystems
-- Correlation and composition heatmaps for microbiome data annotated with plots showing each taxon's prevalence and/or abundance
-- A compact cladogram visualization approach for intuitive comparison of numerous microbe-metadata associations derived from (multivariable) statistical models (taxonomic association trees)
+-   A Shiny app [@shiny] for interactive exploration of microbiota data within R, pairing ordination plots with abundance bar charts
+-   Easy to use functions for generating publication-ready ordination plots with ggplot2 [@wickham_ggplot2_2016], accommodating constrained and partial ordination, bi-plots and tri-plots, and automatic captioning designed to promote methodological transparency and reproducibility
+-   A novel visualization approach pairing ordination plots with circular bar charts (iris plots) for comprehensive, intuitive and compact visualization of the similarity and composition of hundreds of microbial ecosystems
+-   Correlation and composition heatmaps for microbiome data annotated with plots showing each taxon's prevalence and/or abundance
+-   A compact cladogram visualization approach for intuitive comparison of numerous microbe-metadata associations derived from (multivariable) statistical models (taxonomic association trees)
 
 # Statement of need
 
-Modern microbiome research typically involves the use of next-generation-sequencing methods to profile the relative abundance of hundreds of microbial taxa across tens, hundreds or thousands of samples.
[…]
