+++ manuscript/microViz.bib
@@ -105,3 +105,34 @@
   note = {R package version 1.6.1},
   url = {https://CRAN.R-project.org/package=pkgdown},
 }
+
+@article{chen_2012,
+	title = {Associating microbiome composition with environmental covariates using generalized {UniFrac} distances},
+	volume = {28},
+	issn = {1460-2059, 1367-4803},
+	url = {https://academic.oup.com/bioinformatics/article-lookup/doi/10.1093/bioinformatics/bts342},
+	doi = {10.1093/bioinformatics/bts342},
+	language = {en},
+	number = {16},
+	urldate = {2021-05-28},
+	journal = {Bioinformatics},
+	author = {Chen, Jun and Bittinger, Kyle and Charlson, Emily S. and Hoffmann, Christian and Lewis, James and Wu, Gary D. and Collman, Ronald G. and Bushman, Frederic D. and Li, Hongzhe},
+	month = aug,
+	year = {2012},
+	pages = {2106--2113},
+}
+
+@article{gloor_2017,
+	title = {Microbiome {Datasets} {Are} {Compositional}: {And} {This} {Is} {Not} {Optional}},
+	volume = {8},
+	issn = {1664-302X},
+	shorttitle = {Microbiome {Datasets} {Are} {Compositional}},
+	url = {http://journal.frontiersin.org/article/10.3389/fmicb.2017.02224/full},
+	doi = {10.3389/fmicb.2017.02224},
+	urldate = {2021-05-28},
+	journal = {Frontiers in Microbiology},
+	author = {Gloor, Gregory B. and Macklaim, Jean M. and Pawlowsky-Glahn, Vera and Egozcue, Juan J.},
+	month = nov,
+	year = {2017},
+	pages = {2224},
+}
