### 2021-05-28 3f933fb
JOSS manuscript small tweak
