### 2021-05-28 david-barnett/microViz#25 @david-barnett
Thanks for highlighting these issues.

- I was aware that problems arise with the specifying the shape. This error is emitted by the ggiraph package that microViz depends on to provide the plot interactivity. I will try to isolate why this happens for some shape specifications and not others, and see what I can do to avoid this limitation or at least work around it more cleanly.
- Regarding your points about the lasso not always being accurate, I have not seen this, but I will investigate :)
- Some of the options in the "Add:" menu are not expected to work with all ordination plots, and I should make this clearer in the documentation, or perhaps dynamically remove those options when they aren't applicable e.g.
   - taxa loadings cannot be drawn for distance-based ordinations
   - hulls and ellipses cannot be drawn without first mapping colour to a variable, to provide groups
