### 2021-05-30 ef0ecef
fix: ord_explore shape scale limited to avoid ggiraph errors
