+++ R/ord_explore.R
@@ -1,34 +1,44 @@
 #' Interactively explore microbial compositions of ordinated samples
 #'
 #' @description
-#' A Shiny app used like an interactive version of `ord_plot()`.
-#' You can select samples on an ordination plot and view their composition with stacked barplots.
+#' A Shiny app used to create and explore an interactive version of `ord_plot()`.
+#' You can select samples on an ordination plot to view their composition with stacked barplots.
 #'
 #' The `ord_explore()` data argument takes either:
 #'
 #' - the output of `ord_calc()` (i.e. a ps_extra with an ordination)
 #' - a plain phyloseq object: `ord_explore()` will help you build an ordination
 #'
+#' Once the app is running (in your browser), you can:
 #'
-#' Once the app is running (in your browser):
+#' 1. Create/edit the ordination if required
+#'    - look at the R console error messages if your chosen options don't build
+#' 2. Style the ordination plot (e.g. choose dimensions; set colour and size; ...)
+#'    - Taxa loading arrows can be added only to PCA, RDA and CCA plots
+#'    - Convex hulls or ellipses can only be drawn if Colour is set to a variable
+#'    - To track individuals over time with the path plotter, your data MUST already be sorted by time (e.g. with ps_arrange)!
+#' 3. Click on or use the lasso tool to select 2 or more samples to view their compositions
+#'    - By default samples can be selected individually
+#'    - Set the "Select" option to another variable to select by level of that variable
+#' 4. Style the taxonomic compositions barplot
+#'    - The samples are ordered using the distance method
+#'    - The app may lag if you select 100s of samples and ungroup the "other" category
+#'    - To avoid this lag: either reduce the number of taxa or samples, or deselect "Interactive" barplot
+#' 5. Stop the app by clicking the red stop button in the R console
+#'    - Closing the web browser window doesn't stop the app,
+#'   (you can find the app again at the local http address shown in the R console)
+#'    - Don't forget to copy the ordination plot code before you close the app
 #'
-#' 1. edit the ordination if required
-#' 2. style the ordination plot (e.g. choose dimensions; set colour and size; ...)
-#' 3. click or use the lasso tool to select 2 or more samples to view their compositions
-#' 4. style the taxonomic compositions barplot
-#' 5. stop the app by clicking the red stop button in the R console
-#' (closing the web browser window doesn't stop the app)
+#'  See the Details section for some known limitations of the app.
+#'  Please report any other app problems on the microViz GitHub issues page.
 #'
 #' @details
-#' If you get an interactive error like the one below:
+#' Limitations:
 #'
-#' "ids don't have the same length than str (most often, it occurs because of clipping)"
-#'
-#' 1. make your points smaller
-#' 2. set point shape to fixed, or to a variable with fewer categories
-#'
-#' Limitation: When a selection grouping variable is NA for some samples,
-#' that grouping variable cannot be used to select those samples
+#' - If a "Select:" grouping variable is NA for some samples,
+#' then that grouping variable cannot be used to select those samples
+#' - "Shape:" can only be mapped to variables with a maximum of 5 distinct levels,
+#' not including NAs. NAs in the shape variable are shown as hollow circles.
 #'
 #' On some web browsers, e.g. firefox, the numeric inputs' buttons are sometimes
[…]
