+++ NAMESPACE
@@ -42,6 +42,7 @@ export(ps_reorder)
 export(ps_select)
 export(ps_seriate)
 export(samdat_tbl)
+export(scale_shape_girafe_filled)
 export(stat_chull)
 export(tax_agg)
 export(tax_anno)
