### 2021-05-30 c12e01f
ord_explore only offers shape variables that can be drawn without error (max 5 values and 1 NA)
