### 2021-05-30 92d0546
ord_explore drop defunct N distinct slider update upon unmerge, this was never working well
