### 2021-05-30 6e09e3d
ord_explore style: replace sliders with numeric inputs
