+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.6.9002
+Version: 0.7.6.9003
 Authors@R: 
     person(given = "David",
            family = "Barnett",
