### 2021-05-30 david-barnett/microViz#25 @david-barnett
I think I have resolved the issue with the shapes error (as far as is possible) in the latest development version. Since commit: c12e01f19da5b6bf3078950d5e4aa97890d5d124 the user can select only variables with a max of 5 distinct levels (not including NA) for use as the shape mapping. A custom shape palette is now used which avoids using the composite shapes that were used by ggplot2 default and were causing the ggiraph dependency's error:  `Error: ids don't have the same length than str (most often, it occurs because of clipping)`

I have also added a larger description of how to use the app to the ord_explore [docs](https://david-barnett.github.io/microViz/reference/ord_explore.html) that might help somewhat with the other problems. Of course, please do let me know if you still have some issues with ord_explore.

I can't reproduce your experience of inaccurate lasso selection, what OS and web browser are you using? Is it selecting nearby points or completely incorrect points?
Thanks!
