+++ README.md
@@ -78,7 +78,7 @@ if (!requireNamespace("BiocManager", quietly = TRUE)) install.packages("BiocMana
 BiocManager::install(c("phyloseq", "microbiome"))
 
 # To install the latest "released" version of this package
-devtools::install_github("david-barnett/microViz@0.7.6") # check 0.7.6 is the latest release
+devtools::install_github("david-barnett/microViz@0.7.7") # check 0.7.7 is the latest release
 
 # To install the very latest version:
 devtools::install_github("david-barnett/microViz")
@@ -265,8 +265,8 @@ aitchison_perm <- dist_permanova(
   variables = "bmi_group + female"
 )
 #> Dropping samples with missings: 2
-#> 2021-05-20 08:57:15 - Starting PERMANOVA with 99 perms with 1 processes
-#> 2021-05-20 08:57:16 - Finished PERMANOVA
+#> 2021-06-04 09:22:53 - Starting PERMANOVA with 99 perms with 1 processes
+#> 2021-06-04 09:22:53 - Finished PERMANOVA
 # view the permanova results
 perm_get(aitchison_perm) %>% as.data.frame()
 #>            Df   SumOfSqs         R2        F Pr(>F)
@@ -294,8 +294,8 @@ your permanova directly using the ord\_plot function with constraints.
 ``` r
 perm2 <- dist_permanova(data = aitchison_dists, variables = c("weight", "female"), seed = 321)
 #> Dropping samples with missings: 2
-#> 2021-05-20 08:57:16 - Starting PERMANOVA with 999 perms with 1 processes
-#> 2021-05-20 08:57:16 - Finished PERMANOVA
+#> 2021-06-04 09:22:53 - Starting PERMANOVA with 999 perms with 1 processes
+#> 2021-06-04 09:22:54 - Finished PERMANOVA
 perm_get(perm2)
 #> Permutation test for adonis under reduced model
 #> Marginal effects of terms
@@ -398,35 +398,36 @@ sessionInfo()
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
 #> 
 #> other attached packages:
-#> [1] ggplot2_3.3.3   dplyr_1.0.6     phyloseq_1.34.0 microViz_0.7.6  devtools_2.4.0  usethis_2.0.1  
-#> [7] pkgdown_1.6.1  
+#> [1] ggplot2_3.3.3       dplyr_1.0.6         phyloseq_1.34.0     microViz_0.7.6.9002
+#> [5] devtools_2.4.0      usethis_2.0.1       pkgdown_1.6.1      
 #> 
 #> loaded via a namespace (and not attached):
 #>   [1] nlme_3.1-149         matrixStats_0.58.0   fs_1.5.0             RColorBrewer_1.1-2  
 #>   [5] progress_1.2.2       rprojroot_2.0.2      tools_4.0.3          utf8_1.2.1          
-#>   [9] R6_2.5.0             vegan_2.5-7          BiocGenerics_0.36.1  mgcv_1.8-33         
-#>  [13] colorspace_2.0-1     GetoptLong_1.0.5     permute_0.9-5        rhdf5filters_1.2.1  
-#>  [17] ade4_1.7-16          withr_2.4.2          tidyselect_1.1.1     prettyunits_1.1.1   
-#>  [21] processx_3.5.1       compiler_4.0.3       microbiome_1.12.0    cli_2.5.0           
-#>  [25] Biobase_2.50.0       Cairo_1.5-12.2       TSP_1.1-10           desc_1.3.0          
[…]
