+++ _pkgdown.yml
@@ -53,6 +53,7 @@ reference:
   - ord_plot
   - ord_plot_iris
   - ord_explore
+  - scale_shape_girafe_filled
   - ps_extra-accessors
 
 - title: "Taxon models and tree visualisations"
