### 2021-06-08 20ca68b
refactor ord_plot by extracting ord_caption helper function
