### 2021-06-09 32b7c30
ord_plotArrows internal helper and vec_* helpers
