+++ R/ord_plot.R
@@ -163,13 +163,13 @@ ord_plot <-
            axes = 1:2,
            plot_taxa = FALSE,
            tax_vec_length = NA,
-           tax_vec_style_all = list(),
-           tax_vec_style_sel = list(),
+           tax_vec_style_all = vec_tax_all(),
+           tax_vec_style_sel = vec_tax_sel(),
            tax_lab_length = tax_vec_length * 1.1,
            tax_lab_style = list(),
            taxon_renamer = function(x) identity(x),
            constraint_vec_length = NA,
-           constraint_vec_style = list(),
+           constraint_vec_style = vec_constraint(),
            constraint_lab_length = constraint_vec_length * 1.1,
            constraint_lab_style = list(),
            var_renamer = function(x) identity(x),
@@ -345,30 +345,18 @@ ord_plot <-
         }
 
         # (semi-transparent) lines for all taxa
-        tax_vec_all_args <- list(
-          data = speciesScoresDf * tax_vec_length,
-          mapping = ggplot2::aes_string(
-            xend = axesNames[1], yend = axesNames[2], x = 0, y = 0
-          ),
-          size = 0.5, alpha = 0.25
+        p <- ord_arrows(
+          p = p, data = speciesScoresDf * tax_vec_length,
+          axesNames = axesNames, styleList = tax_vec_style_all,
+          defaultStyles = vec_tax_all()
         )
-        tax_vec_all_args[names(tax_vec_style_all)] <- tax_vec_style_all
-        p <- p + do.call(what = ggplot2::geom_segment, args = tax_vec_all_args)
 
         # (opaque) lines for selected taxa
-        tax_vec_sel_args <- list(
-          data = selectSpeciesScoresDf * tax_vec_length,
-          mapping = ggplot2::aes_string(
-            xend = axesNames[1], yend = axesNames[2], x = 0, y = 0
-          ),
-          lineend = "round", linejoin = "mitre",
-          arrow = grid::arrow(
-            length = grid::unit(0.005, "npc"), type = "closed"
-          ),
-          size = 0.5
+        p <- ord_arrows(
+          p = p, data = selectSpeciesScoresDf * tax_vec_length,
+          axesNames = axesNames, styleList = tax_vec_style_sel,
+          defaultStyles = vec_tax_sel()
         )
-        tax_vec_sel_args[names(tax_vec_style_sel)] <- tax_vec_style_sel
-        p <- p + do.call(what = ggplot2::geom_segment, args = tax_vec_sel_args)
 
         # selected taxa labels
         tax_lab_args <- list(
@@ -390,20 +378,11 @@ ord_plot <-
       }
       # draw vector segments at length set by constraint_vec_length argument
       # (proportion of original length)
-      constraint_vec_args <- list(
-        mapping = ggplot2::aes_string(
-          xend = axesNames[1], yend = axesNames[2], x = 0, y = 0
-        ),
-        alpha = 0.8, colour = "brown",
-        lineend = "round", linejoin = "mitre",
-        arrow = grid::arrow(
-          length = grid::unit(0.005, "npc"), type = "closed"
-        ),
-        data = constraintDf * constraint_vec_length
+      p <- ord_arrows(
+        p = p, data = constraintDf * constraint_vec_length,
+        axesNames = axesNames, styleList = constraint_vec_style,
+        defaultStyles = vec_constraint()
       )
-      constraint_vec_args[names(constraint_vec_style)] <- constraint_vec_style
-      p <- p +
-        do.call(what = ggplot2::geom_segment, args = constraint_vec_args)
 
       # draw vector tip labels at length set by constraint_lab_length argument
[…]
