+++ NEWS.md
@@ -1,5 +1,7 @@
 # microViz (development version)
 
+- `ord_plot` gains `vec_*` helper functions for generating lists for styling taxa and constraint vectors/arrows (`vec_constraint`, `vec_tax_all` and `vec_tax_sel`)
+
 # microViz 0.7.7
 
 - `ord_explore` shapes selection bug fixed by limiting to 5 shapes returned by new `scale_shape_girafe_filled` function
