+++ man/ord_plot.Rd
@@ -9,13 +9,13 @@ ord_plot(
   axes = 1:2,
   plot_taxa = FALSE,
   tax_vec_length = NA,
-  tax_vec_style_all = list(),
-  tax_vec_style_sel = list(),
+  tax_vec_style_all = vec_tax_all(),
+  tax_vec_style_sel = vec_tax_sel(),
   tax_lab_length = tax_vec_length * 1.1,
   tax_lab_style = list(),
   taxon_renamer = function(x) identity(x),
   constraint_vec_length = NA,
-  constraint_vec_style = list(),
+  constraint_vec_style = vec_constraint(),
   constraint_lab_length = constraint_vec_length * 1.1,
   constraint_lab_style = list(),
   var_renamer = function(x) identity(x),
