+++ man/ord_plotArrows.Rd
@@ -1,0 +1,53 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ord_plotArrows.R
+\name{vec_constraint}
+\alias{vec_constraint}
+\alias{vec_tax_sel}
+\alias{vec_tax_all}
+\title{create ordination plot vector styling lists}
+\usage{
+vec_constraint(
+  size = 1,
+  alpha = 0.8,
+  colour = "brown",
+  arrow = grid::arrow(length = grid::unit(0.005, units = "npc"), type = "closed", angle
+    = 30),
+  lineend = "round",
+  linejoin = "mitre",
+  ...
+)
+
+vec_tax_sel(
+  size = 0.5,
+  alpha = 1,
+  colour = "black",
+  arrow = grid::arrow(length = grid::unit(0.005, units = "npc"), type = "closed", angle
+    = 30),
+  lineend = "round",
+  linejoin = "mitre",
+  ...
+)
+
+vec_tax_all(size = 0.5, alpha = 0.25, arrow = NULL, ...)
+}
+\arguments{
+\item{size}{width of vector}
+
+\item{alpha}{opacity of vector}
+
+\item{colour}{colour of vector}
+
+\item{arrow}{arrow style specified with grid::arrow() or NULL for no arrow}
+
+\item{lineend}{Line end style (round, butt, square).}
+
+\item{linejoin}{Line join style (round, mitre, bevel).}
+
+\item{...}{further arguments passed to geom_segment}
+}
+\value{
+list
+}
+\description{
+create ordination plot vector styling lists
+}
