+++ NAMESPACE
@@ -64,5 +64,8 @@ export(taxatree_nodes)
 export(taxatree_plotkey)
 export(taxatree_plots)
 export(var_anno)
+export(vec_constraint)
+export(vec_tax_all)
+export(vec_tax_sel)
 importFrom(dplyr,"%>%")
 importFrom(rlang,.data)
