### 2021-06-09 c91b435
ord_calc refactor extract internal check helpers
