+++ R/ord_calc.R
@@ -92,123 +92,54 @@ ord_calc <- function(data,
                      scale_cc = TRUE,
                      verbose = TRUE,
                      ...) {
-  stopifnot(method %in% c(
-    "auto", "PCoA", "PCA", "CCA", "RDA", "CAP", "NMDS", "MDS", "DCA", "DPCoA"
-  ))
   stopifnot(inherits(scale_cc, "logical"))
-  if (!isFALSE(verbose) && method %in% c("DCA", "DPCoA")) {
-    warning(
-      "Neither DCA nor DPCoA have been tested, so may not work as expected..."
-    )
-  }
 
   # identify if constrained or conditioned
-  if (identical(constraints, NULL) || identical(constraints, 1)) {
-    constraints <- 1
-    constrained <- FALSE
-  } else {
-    if (!inherits(constraints, "character")) {
-      stop("constraints must be NULL or a character vector")
-    }
-    constrained <- TRUE
-  }
-  conditioned <- !identical(conditions, NULL)
-  if (conditioned && !inherits(conditions, "character")) {
-    stop("conditions must be NULL or a character vector")
-  }
+  if (identical(constraints, NULL)) constraints <- 1
+  constrained <- ordCheckConstraints(constraints)
+  conditioned <- ordCheckConditions(conditions)
 
-  # if input is a phyloseq, convert to ps_extra
-  if (inherits(data, "phyloseq")) data <- new_ps_extra(ps = data)
-  # check class of data is (or at least now is) ps_extra
-  if (!inherits(data, "ps_extra")) {
-    stop(
-      "data should be ps_extra list output of dist_calc or tax_transform, ",
-      "or a phyloseq\n", "data is class: ", class(data)
-    )
-  }
+  # check class of data is ps_extra and if phyloseq, convert to ps_extra
+  data <- ordCheckData(data)
 
   # get data elements
   ps <- ps_get(data)
   distMat <- dist_get(data)
   info <- info_get(data)
 
-  # set method automatically if auto given
-  if (identical(method, "auto")) {
-    if (identical(distMat, NULL)) {
-      method <- if (!constrained) "PCA" else "RDA"
-    } else {
-      method <- if (!constrained) "PCoA" else "CAP"
-    }
-  }
-  # check compatibility of input
   if (!identical(ord_get(data), NULL) && !isFALSE(verbose)) {
     warning(
       "You already calculated an ordination (with ",
       info[["ordMethod"]], ") --> overwriting"
     )
   }
-  if (!identical(distMat, NULL) &&
-    method %in% c("PCA", "RDA", "CCA") &&
-    !isFALSE(verbose)) {
-    warning(
-      "Distance matrix is not used for ", method,
-      "! Removing distance matrix. ",
-      "Did you mean to use PCoA or CAP? (or use method = auto)"
-    )
-    distMat <- NULL # so output ps_extra will get NULL distance matrix
-  }
-  if (identical(distMat, NULL) && method %in% c("CAP", "PCoA", "MDS", "NMDS")) {
-    stop(
-      "Distance matrix missing! Use dist_calc() ",
-      "before using ord_calc() with this method: ", method
-    )
-  }
-  if (constrained && method %in% c("PCA", "PCoA", "NMDS", "MDS")) {
-    stop(
-      method, " cannot use constraints, did you mean RDA, CAP or CCA. ",
-      "Alternatively use method = auto (the default)"
-    )
-  }
+
+  # check method is valid option, and guess appropriate method if "auto"
+  method <- ordCheckMethod(
+    method = method, con = constrained, distMat = distMat, verbose = verbose
+  )
+  # check ordination method compatible with distance availability
[…]
