### 2021-06-12 #3201 @david-barnett
>
>
> Hi @david-barnett, just wanted to let you know that I have posted a couple of issues related to questions I had regarding the statement of need ([david-barnett/microViz#26](https://github.com/david-barnett/microViz/issues/26)) and the functionality of the RShiny app ([david-barnett/microViz#25](https://github.com/david-barnett/microViz/issues/25)). Thanks!

Thanks for raising these issues marypiper, I have taken actions to address your points, and I included descriptions of the changes I have made directly on the relevant issues. Do let me know if there is more to be done. Thanks!
