### 2021-06-12 #3201 @marypiper
@lpantano, I have completed my review of microViz. I initially had issues testing out the interactive component of microViz and had questions regarding the novelty of the functions. However, @david-barnett addressed these, fixing the interactive visualizations so that they now function appropriately and he has provided additional documentation for users. David has provided the updated reference for the microbiome citation, and he has provided more explanations regarding the usefulness of his contributions. While the methods proposed aren't in and of themselves novel, he has brought together useful functions from quite a few microbiome analysis packages and provided wrappers to make them easier to use, especially for a more novice R user working to perform visualizations of their microbiome data. David has created useful visualizations that are aesthetically pleasing and often not available in the main packages for microbiome data analysis/visualization, in addition to providing the ordination plots as an interactive RShiny package.
