+++ vignettes/articles/videos.Rmd
@@ -1,0 +1,12 @@
+---
+title: "Videos"
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+<div style="padding:56.25% 0 0 0;position:relative;"><iframe src="https://player.vimeo.com/video/559627248?badge=0&amp;autopause=0&amp;player_id=0&amp;app_id=58479" frameborder="0" allow="autoplay; fullscreen; picture-in-picture" allowfullscreen style="position:absolute;top:0;left:0;width:100%;height:100%;" title="microViz 10-minute intro/demo"></iframe></div><script src="https://player.vimeo.com/api/player.js"></script>
