+++ _pkgdown.yml
@@ -85,3 +85,14 @@ reference:
   - phyloseq_validate
 
 
+articles:
+  - title: Tutorials
+    navbar: Tutorials
+    contents:
+    - articles/phyloseq
+    - articles/tax-fixing
+    - articles/ordination
+    - articles/interactive-ordination
+    - articles/compositions
+    - articles/videos
+
