+++ pkgdown.yml
@@ -2,13 +2,14 @@ pandoc: 2.7.3
 pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
-  articles/Interactive-Ordination: Interactive-Ordination.html
-  articles/Visualising-compositions: Visualising-compositions.html
+  articles/compositions: compositions.html
+  articles/interactive-ordination: interactive-ordination.html
   articles/microViz: microViz.html
+  articles/ordination: ordination.html
   articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
-  ordination: ordination.html
-last_built: 2021-06-04T07:30Z
+  articles/videos: videos.html
+last_built: 2021-06-14T20:44Z
 urls:
   reference: https://david-barnett.github.io/microViz//reference
   article: https://david-barnett.github.io/microViz//articles
