+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.7.9002
+Version: 0.7.8
 Authors@R: 
     person(given = "David",
            family = "Barnett",
