+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   articles/phyloseq: phyloseq.html
   articles/tax-fixing: tax-fixing.html
   articles/videos: videos.html
-last_built: 2021-06-14T20:44Z
+last_built: 2021-06-21T19:17Z
 urls:
   reference: https://david-barnett.github.io/microViz//reference
   article: https://david-barnett.github.io/microViz//articles
