### 2021-06-25 a4cf691
ord_explore barplot legend fix due to ggplot2 3.3.4 bug https://github.com/tidyverse/ggplot2/issues/4511
