+++ NEWS.md
@@ -1,3 +1,7 @@
+# microViz (development version)
+
+- fix (hopefully temporary) for incorrect barplot legend in ord_explore caused by bug introduced by ggplot2 version 3.3.4 noted at https://github.com/tidyverse/ggplot2/issues/4511
+
 # microViz 0.7.8
 
 - `ord_plot` gains `vec_*` helper functions for generating lists for styling taxa and constraint vectors/arrows (`vec_constraint`, `vec_tax_all` and `vec_tax_sel`)
