### 2021-06-26 #3201 @yoonjeongcha
@lpantano, I am now also done with my review. The visualization and interactive exploration were easy to use and helpful. The code was clean and well documented, and the vignettes had great examples. The tutorials and video was also well made. Overall I thought it was a great package for those who want to explore and visualize microbiome sequencing data in a single package.
