+++ README.Rmd
@@ -71,7 +71,7 @@ if (!requireNamespace("BiocManager", quietly = TRUE)) install.packages("BiocMana
 BiocManager::install(c("phyloseq", "microbiome"))
 
 # To install the latest "released" version of this package
-devtools::install_github("david-barnett/microViz@0.7.8") # check 0.7.8 is the latest release
+devtools::install_github("david-barnett/microViz@0.7.9") # check 0.7.9 is the latest release
 
 # To install the very latest version:
 devtools::install_github("david-barnett/microViz")
