+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: microViz
 Title: Microbiome Data Analysis and Visualization
-Version: 0.7.8.9000
+Version: 0.7.9
 Authors@R: 
     person(given = "David",
            family = "Barnett",
