### 2021-06-30 #3201 @david-barnett
Great! 🙂 and thank you all so much for volunteering your time

@lpantano here is the latest tagged release and the Zenodo archive, I hope this is all in order
https://github.com/david-barnett/microViz/releases/tag/0.7.9
https://doi.org/10.5281/zenodo.5048013
